% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk__bind_mouse_keys.R
\name{bind_mouse_keys}
\alias{bind_mouse_keys}
\title{Bind mouse click/release events to a TclTk widget}
\usage{
bind_mouse_keys(obj, envir = parent.frame())
}
\arguments{
\item{obj}{A TclTk widget (unquoted name).}

\item{envir}{Environment for evaluation.}
}
\description{
Bind mouse click/release events to a TclTk widget
}
\keyword{internal}
