% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_elements.R
\name{box_ask_to_replace}
\alias{box_ask_to_replace}
\alias{replace_object}
\alias{replace_variable}
\title{Message box, which asks if object should be replaced}
\usage{
box_ask_to_replace(name, type = gettextRcmdr("Object"))

replace_object(name, type = gettextRcmdr("Object"))

replace_variable(name, type = gettextRcmdr("Variable"))
}
\arguments{
\item{name}{The name of the object.}

\item{type}{The type of the object.}
}
\value{
Logical value: TRUE -- to replace or FALSE -- not to replace.
}
\description{
Message box, which asks if object should be replaced
}
\keyword{internal}
