% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_checkboxes.R
\name{bs_checkboxes}
\alias{bs_checkboxes}
\title{Checkboxes widget.}
\usage{
bs_checkboxes(
  parent = top,
  boxes,
  labels = NULL,
  title = NULL,
  values = NULL,
  default_value = "0",
  commands = list(),
  default_command = function() {
 },
  tips = list(),
  default_tip = "",
  border = FALSE,
  layout = c("vertical", "horizontal"),
  sticky_buttons = "w",
  sticky_title = "w"
)
}
\arguments{
\item{parent}{Parent frame.}

\item{boxes}{(vector of strings) Variable names for each checkbox.}

\item{labels}{(vector of strings) Labels for each checkbox.}

\item{title}{Title for the set of checkboxes.}

\item{values}{A vector of values ("0" or "1").
Overrides the `default_value`.}

\item{default_value}{(string) Default value ("0" or "1").}

\item{commands}{A named list of commands (functions) for checkbox.
The names must match the values of "boxes".}

\item{default_command}{(function) A default command.}

\item{tips}{A named list of strings to be used as tips for checkbox.
The names must match the values of "boxes".}

\item{default_tip}{(string) a default tip.}

\item{border}{(logical) Flag if the frame should have a border.}

\item{layout}{(string) One of "vertical" (default) and "horizontal".}

\item{sticky_buttons}{(string) `sticky` option for buttons.}

\item{sticky_title}{(string) `sticky` option for title
(if no border is used).}
}
\value{
A named list with fields `frame` (frame with the checkboxes),
 `var` (tcl/tk variables for each box),
 and `obj` (tcl/tk objects for each box).
}
\description{
Checkboxes widget with command functions and tips when mouse is over the box.
}
\examples{
\dontrun{

library(RcmdrPlugin.biostat)

top <- tcltk::tktoplevel()

boxes_1 <- bs_checkboxes(top, c("A", "B", "C"))
tcltk::tkgrid(boxes_1$frame)


top <- tcltk::tktoplevel()
boxes_2 <- bs_checkboxes(top, boxes = c("A", "B", "C"), border = TRUE)
tcltk::tkgrid(boxes_2$frame)


top <- tcltk::tktoplevel()
boxes_3 <- bs_checkboxes(top, c("A", "B", "C"),
  layout = "h", title = "Buttons")
tcltk::tkgrid(boxes_3$frame)


set_values(boxes_3, B = TRUE, C = TRUE)

new_vals <- c(A = TRUE, B = FALSE)
set_values(boxes_3, new_vals)


get_values(boxes_3)
get_values(boxes_3, "B")
get_values(boxes_3, simplify = FALSE)
}
}
