% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk__bs_combobox.R
\name{bs_combobox}
\alias{bs_combobox}
\title{TclTk ComboBox widget constructor}
\usage{
bs_combobox(
  parent = top,
  values = NULL,
  state = c("readonly", "normal", "disabled"),
  value = NULL,
  selection = NULL,
  combobox_sticky = "nw",
  on_select = do_nothing,
  postcommand = do_nothing,
  on_click = do_nothing,
  on_double_click = do_nothing,
  on_triple_click = do_nothing,
  on_release = do_nothing,
  on_click_3 = do_nothing,
  on_double_click_3 = do_nothing,
  on_triple_click_3 = do_nothing,
  on_release_3 = do_nothing,
  tip =
    gettext_bs("For quicker search, press the \\nfirst letter of variable name\\non your keyboard."),
  width = 20,
  height = 8,
  on_keyboard = c("select", "ignore"),
  export = "FALSE",
  label = "",
  label_position = c("left", "above", "right", "none"),
  label_color = getRcmdr("title.color"),
  padx = 0,
  pady = 0,
  sticky = "w",
  sticky_label = sticky,
  sticky_text = sticky,
  main_frame = tk2frame(parent),
  combobox_frame = tk2frame(main_frame),
  label_frame = tk2frame(main_frame),
  label_tip = "",
  ...
)
}
\arguments{
\item{parent}{Parent TclTk frame.}

\item{values}{Character vector of choices.}

\item{state}{Widget state.}

\item{selection}{Initial selection index.}

\item{combobox_sticky}{Sticky option for the combobox.}

\item{export}{Logical.}

\item{title}{Label text.}

\item{title_sticky}{Sticky option for the title.}
}
\description{
Creates a labeled combobox widget.
}
\keyword{internal}
