% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_listbox_y_gr.R
\name{bs_listbox_y_gr}
\alias{bs_listbox_y_gr}
\title{Widget: two listboxes with checkbox}
\usage{
bs_listbox_y_gr(
  parent = top,
  y_title = title_var_1,
  y_var_type = NA_character_,
  y_vars = NULL,
  y_initial = NULL,
  y_select_mode = "single",
  y_tip = "",
  y_params = list(),
  gr_title = title_gr_0_n,
  gr_var_type = NA_character_,
  gr_vars = NULL,
  gr_initial = NULL,
  gr_select_mode = "multiple",
  gr_on_select = do_nothing,
  gr_tip = "",
  gr_params = list(),
  ch_label = "Use groups",
  ch_initial = "0",
  ch_command = do_nothing,
  ch_tip = "",
  ch_params = list(),
  list_height = 7,
  list_width = c(25, Inf),
  add_to_grid = TRUE
)
}
\description{
Widget: two listboxes with checkbox
}
\examples{
\dontrun{

# Active dataset must be selected
top <- tktoplevel()
lb <- bs_listbox_y_gr(top)
tkgrid(lb$frame)
}
}
\keyword{internal}
