% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{characterP}
\alias{characterP}
\title{Does data contain characters?}
\usage{
characterP(n = 1)
}
\arguments{
\item{n}{Minimum number of character variables}
}
\description{
Return TRUE, if at least n character variables exist in the active dataset.
}
\keyword{internal}
