% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aaa.r
\docType{data}
\name{dataContour}
\alias{dataContour}
\title{A Dummy Data for Contour Plots}
\format{A data frame with 10000 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{a x-axis variable}
    \item{\code{y}}{a y-axis variable}
    \item{\code{z}}{density}
  }}
\description{
A Dummy Data for Contour Plots
}
\examples{
# This data is generated by the code below:
# set.seed(5435678)
# dataContour <- data.frame(
#   x = (zx <- (rep(1:100, 100) - 5) / 2),
#   y = (zy <- (rep(1:100, each=100) - 5) / 2),
#   z = (z  <- zx*zy)
# )
# save(dataContour, file = "dataContour.RData")
# try(data(dataContour, package = "RcmdrPlugin.KMggplot2"))
}
\seealso{
contour-class
}
\keyword{data}
