% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aaa.r
\docType{data}
\name{dataLine}
\alias{dataLine}
\title{A Dummy Data for Line Charts}
\format{A data frame with 144 observations on the following 5 variables.
  \describe{
    \item{\code{date}}{date}
    \item{\code{group}}{a measurement group (has three groups "T", "H", and "L")}
    \item{\code{marker}}{a biomarker level (has two groups "High", and "Low")}
    \item{\code{sex}}{sex}
    \item{\code{y}}{a measurement variable}
  }}
\description{
A Dummy Data for Line Charts
}
\examples{
# This data is generated by the code below:
# set.seed(5435678)
# date <- rep(seq(Sys.Date(), len = 12, by = "1 month"), each = 12)
# group <- rep(c("T", "H", "L"), 12, each = 4)
# marker <- factor(rep(c("High", "Low"), 36, each = 2))
# sex <- factor(rep(c("M", "F"), 72))
# mu <- rep(c(200, 50, 150), 12, each = 4) + 10 * as.numeric(sex) + 30 * as.numeric(marker)
# y <- rnorm(144, mu, 25)
# dataLine <- data.frame(date, group, marker, sex, y)
# save(dataLine, file = "dataLine.RData")
# try(data(dataLine, package = "RcmdrPlugin.KMggplot2"))
}
\seealso{
line-class
}
\keyword{data}
