% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezr-functions.R
\name{ezr-functions}
\alias{ezr-functions}
\alias{StatMedCorrelation}
\alias{StatMedSpearman}
\alias{StatMedSingleSampleTTest}
\alias{StatMedWilSign}
\alias{StatMedPairedTtest}
\alias{StatMedFriedman}
\alias{StatMedTtest}
\alias{StatMedMannW}
\alias{StatMedANOVA}
\alias{StatMedKruWalli}
\alias{StatMedJT}
\alias{StatMedMultiANOVA}
\alias{StatMedANCOVA}
\alias{StatMedBartlett}
\alias{StatMedFTest}
\title{\pkg{RcmdrPlugin.EZR} functions}
\usage{
StatMedCorrelation()

StatMedSpearman()

StatMedSingleSampleTTest()

StatMedWilSign()

StatMedPairedTtest()

StatMedFriedman()

StatMedTtest()

StatMedMannW()

StatMedANOVA()

StatMedKruWalli()

StatMedJT()

StatMedMultiANOVA()

StatMedANCOVA()

StatMedBartlett()

StatMedFTest()
}
\description{
Functions imported from package \pkg{RcmdrPlugin.EZR}.
}
