% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{factors_strict_P}
\alias{factors_strict_P}
\title{Does active dataset contain true factors?}
\usage{
factors_strict_P(n = 1)
}
\arguments{
\item{n}{Minimum number of logical variables}
}
\description{
Return TRUE, if at least n factor variables exist in the active dataset.
Function `Factors()` list factor-like data including characters and logicals.
}
\keyword{internal}
