% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{first_class_is_dataframeP}
\alias{first_class_is_dataframeP}
\title{Is the first class "data.frame"}
\usage{
first_class_is_dataframeP()
}
\description{
Check if the first class of active Rcmdr dataset is "data.frame"
}
\keyword{internal}
