% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{first_class_isP}
\alias{first_class_isP}
\title{[!] Is the first class the same as in brackets?}
\usage{
first_class_isP(df_class)
}
\description{
Check if the first class of active Rcmdr dataset is the same as determined.
}
\keyword{internal}
