% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{get_active_ds}
\alias{get_active_ds}
\title{Get contents of active dataset.}
\usage{
get_active_ds(fail = TRUE)
}
\arguments{
\item{fail}{(logical) When no active dataset is selected:
if \code{TRUE}, the function fails,
if \code{FALSE}, then \code{NULL} is returned.}
}
\value{
The contents of active dataset or \code{NULL}.
}
\description{
Get contents of active dataset.
}
