% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu-import.R
\name{get_ds_info}
\alias{get_ds_info}
\title{Get information about dataset}
\usage{
get_ds_info(str)
}
\arguments{
\item{str}{sting with datasets name, e.g, "datasets::iris"}
}
\value{
Dataframe with size, class and some other information about a dataset.
}
\description{
Get information about dataset
}
\examples{
get_ds_info("datasets::iris")
}
