% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu-import.R
\name{get_ds_info_2}
\alias{get_ds_info_2}
\title{Get information about dataset and its variables}
\usage{
get_ds_info_2(str)
}
\arguments{
\item{str}{sting with datasets name, e.g, "datasets::iris"}
}
\value{
Dataframe with size and variable type frequency
}
\description{
Get information about dataset and its variables
}
\examples{
get_ds_info_2("datasets::iris")
}
