% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu-import.R
\name{get_n_datasets}
\alias{get_n_datasets}
\title{Get number of datasets in a package}
\usage{
get_n_datasets(package)
}
\arguments{
\item{package}{String with valid name of inst6alled package}
}
\value{
Number with number of datasets in a package
}
\description{
Get number of datasets in a package
}
\examples{
get_n_datasets("datasets")
get_n_datasets("purrr")
}
