% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu-import.R
\name{get_obj_dims}
\alias{get_obj_dims}
\title{Get either dimensions or length of a dataset}
\usage{
get_obj_dims(obj)
}
\arguments{
\item{obj}{An R object (dataset).}
}
\value{
A string wiht dimensions of matrix-like or array-like oblect and
        length of other objects.
}
\description{
Get either dimensions or length of a dataset
}
\examples{
get_obj_dims(iris)
}
