% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-checkboxes.R
\docType{methods}
\name{front,checkboxes-method}
\alias{front,checkboxes-method}
\title{The \code{front} Method for \code{checkboxes} Subclass}
\usage{
\S4method{front}{checkboxes}(top, initValues, labels, title = "", right.buttons = FALSE)
}
\arguments{
\item{top}{\code{tkwin} class object; top of widget window.}

\item{initValues}{List of integer; initialization values of check boxes.}

\item{labels}{List of character; values of check boxes labels.}

\item{title}{Character; the title of the check box frame.}

\item{right.buttons}{Boolean; whether check boxes are right aligned or not.}
}
\description{
The \code{front} Method for \code{checkboxes} Subclass
}
\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}
