% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-checkboxes.R
\docType{class}
\name{checkboxes-class}
\alias{checkboxes-class}
\alias{checkboxes}
\title{Check Boxes Subclass}
\description{
\code{checkboxes} class is a subclass for GUI check box frame.
}
\details{
This class is a subclass which make GUI check box frame.
}

\section{Fields}{

\describe{
\item{\code{frame}: }{\code{tkwin} class object; parent of widget window.} 
\item{\code{length}: }{Integer; number of grids.} 
\item{\code{back_list}: }{List of \code{tkwin} class object; list of grids.} 
\item{\code{value}: }{List of \code{tclVar} class object; values of the check box frame.} 
\item{\code{cbcheckboxes}: }{\code{tkwin} class object; the check box frame.} 
}
}

\section{Contains}{

\code{gparts_base}
}

\section{Methods}{

\describe{
\item{\code{back(perline = 3)}: }{\code{back} method for \code{gparts_base} class.}
\item{\code{front(top, initValues, labels, title = "", right.buttons = FALSE)}: }{\code{front} method for \code{checkboxes} subclass.}
}
}

\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}
