% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-aaa.R
\docType{methods}
\name{back,gparts_base-method}
\alias{back,gparts_base-method}
\title{The \code{back} Method for \code{gparts_base} Class}
\usage{
\S4method{back}{gparts_base}(perline = 3)
}
\arguments{
\item{perline}{Integer; controls the number of per-line to GUI grids.}
}
\description{
\code{back} method places GUI grids for R-Commander windows.
}
\seealso{
Other guiparts: \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}

