% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-aaa.R
\docType{class}
\name{gparts_base-class}
\alias{gparts_base-class}
\alias{gparts_base}
\title{Base Class for GUI Frames}
\description{
\code{gparts_base} class is the base class for GUI frames.
}
\details{
This class is a framework for implementing subclasses which make GUI frames.
}

\section{Fields}{

\describe{
\item{\code{frame}: }{\code{tkwin} class object; parent of widget window.} 
\item{\code{length}: }{Integer; number of grids.} 
\item{\code{back_list}: }{List of \code{tkwin} class object; list of grids.} 
}
}

\section{Contains}{

NULL
}

\section{Methods}{

\describe{
\item{\code{back(perline = 3)}: }{\code{back} method for \code{gparts_base} class.}
}
}

\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}
