% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_guiparts-radioboxes.R
\docType{methods}
\name{front,radioboxes-method}
\alias{front,radioboxes-method}
\title{The \code{front} Method for \code{radioboxes} Subclass}
\usage{
\S4method{front}{radioboxes}(top, labels, title = "", initValue = 1, right.buttons = FALSE)
}
\arguments{
\item{top}{\code{tkwin} class object; top of widget window.}

\item{labels}{List of character; values of radio boxes labels.}

\item{title}{Character; the title of the radio box frame.}

\item{initValue}{Integer; the initialization value of the radio box frame.}

\item{right.buttons}{Boolean; whether radio boxes are right aligned or not.}
}
\description{
The \code{front} Method for \code{radioboxes} Subclass
}
\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}
