% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-radioboxes.r
\docType{class}
\name{radioboxes-class}
\alias{radioboxes}
\alias{radioboxes-class}
\title{Radio Boxes Subclass}
\description{
\code{radioboxes} class is a subclass for GUI radio box frame.
}
\details{
This class is a subclass which make GUI radio box frame.
}
\section{Fields}{

\describe{
\item{\code{frame}: }{\code{tkwin} class object; parent of widget window.} 
\item{\code{length}: }{Integer; number of grids.} 
\item{\code{back_list}: }{List of \code{tkwin} class object; list of grids.} 
\item{\code{value}: }{\code{tclVar} class object; the value of the radio box frame.} 
\item{\code{rbradioboxes}: }{\code{tkwin} class object; the radio box frame.} 
}
}

\section{Contains}{

\code{gparts_base}
}

\section{Methods}{

\describe{
\item{\code{back(perline = 3)}: }{\code{back} method for \code{gparts_base} class.}
\item{\code{front(top, labels, title = "", initValue = 1, right.buttons = FALSE)}: }{\code{front} method for \code{radioboxes} subclass.}
}
}
\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}

