% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-textfield.R
\docType{class}
\name{textfield-class}
\alias{textfield}
\alias{textfield-class}
\title{Text Field Subclass}
\description{
\code{textfield} class is a subclass for GUI text field frame.
}
\details{
This class is a subclass which make GUI text field frame.
}
\section{Fields}{

\describe{
\item{\code{frame}: }{\code{tkwin} class object; parent of widget window.}
\item{\code{length}: }{Integer; number of grids.}
\item{\code{back_list}: }{List of \code{tkwin} class object; list of grids.}
\item{\code{value}: }{\code{tclVar} class object; the value of the text field frame.}
}
}

\section{Contains}{

\code{gparts_base}
}

\section{Methods}{

\describe{
\item{\code{back(perline = 3)}: }{\code{back} method for \code{gparts_base} class.}
\item{\code{front(top, initialValue, boxwidth, title)}: }{\code{front} method for \code{textfield} subclass.}
}
}
\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}

