% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-textfields.R
\docType{methods}
\name{front,textfields-method}
\alias{front,textfields-method}
\title{The \code{front} Method for \code{textfields} Subclass}
\usage{
\S4method{front}{textfields}(top, initValues, titles, boxwidth = "20")
}
\arguments{
\item{top}{\code{tkwin} class object; top of widget window.}

\item{initValues}{List of character; initialization values of text fields.}

\item{titles}{List of character; titles of text fields.}

\item{boxwidth}{Character; size of text fields.}
}
\description{
The \code{front} Method for \code{textfields} Subclass
}
\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}
