% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-textfields.r
\docType{class}
\name{textfields-class}
\alias{textfields}
\alias{textfields-class}
\title{Multiple Text Fields Subclass}
\description{
\code{textfields} class is a subclass for GUI multiple text fileds frame.
}
\details{
This class is a subclass which make GUI multiple text fileds frame.
}
\section{Fields}{

\describe{
\item{\code{frame}: }{\code{tkwin} class object; parent of widget window.}
\item{\code{length}: }{Integer; number of grids.}
\item{\code{back_list}: }{List of \code{tkwin} class object; list of grids.}
\item{\code{fields}: }{List of \code{textfield} class object; list of \code{textfield}.}
}
}

\section{Contains}{

\code{gparts_base}
}

\section{Methods}{

\describe{
\item{\code{back(perline = 3)}: }{\code{back} method for \code{gparts_base} class.}
\item{\code{front(top, initValues, titles, boxwidth = "20")}: }{\code{front} method for \code{textfields} subclass.}
}
}
\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{front,variableboxes-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}

