% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guiparts-variableboxes.R
\docType{methods}
\name{front,variableboxes-method}
\alias{front,variableboxes-method}
\title{The \code{front} Method for \code{variableboxes} Subclass}
\usage{
\S4method{front}{variableboxes}(top, types, titles, modes = "default")
}
\arguments{
\item{top}{\code{tkwin} class object; top of widget window.}

\item{types}{List of character; types of variableLists.}

\item{titles}{List of character; titles of variable box frame.}

\item{modes}{List of character; select modes.}
}
\description{
The \code{front} Method for \code{variableboxes} Subclass
}
\seealso{
Other guiparts: \code{\link{back,gparts_base-method}},
  \code{\link{checkboxes-class}},
  \code{\link{front,checkboxes-method}},
  \code{\link{front,radioboxes-method}},
  \code{\link{front,textfield-method}},
  \code{\link{front,textfields-method}},
  \code{\link{front,toolbox-method}},
  \code{\link{gparts_base-class}},
  \code{\link{radioboxes-class}},
  \code{\link{textfield-class}},
  \code{\link{textfields-class}},
  \code{\link{toolbox-class}},
  \code{\link{variableboxes-class}}
}
\keyword{hplot}

