% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdr__initialize_dialog.R
\name{initialize_dialog}
\alias{initialize_dialog}
\title{Initialize a TclTk dialog window (Rcmdr-based macro)}
\usage{
initialize_dialog(
  window = top,
  title = "",
  offset = 10,
  preventCrisp = stop("preventCrisp not supplied"),
  use.tabs = FALSE,
  notebook = notebook,
  tabs = c("dataTab", "optionsTab"),
  suppress.window.resize.buttons = TRUE,
  parent = CommanderWindow()
)
}
\description{
Based on function from Rcmdr v 2.5-1.
}
\keyword{internal}
