% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{is_file_writable}
\alias{is_file_writable}
\title{Check if file is writable.}
\usage{
is_file_writable(file = "")
}
\arguments{
\item{file}{(character) Path to file.}
}
\value{
Logical. If \code{FALSE} it means that file is busy, locked or open.
}
\description{
Check if file is writable.
}
