% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{is_rstudio}
\alias{is_rstudio}
\title{Is R Session in RStudio?}
\usage{
is_rstudio()
}
\value{
Logical value indicating if GUI/IDE is RStudio.
}
\description{
Is R Session in RStudio?
}
\keyword{internal}
