% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{list_objects_of_class}
\alias{list_objects_of_class}
\title{List objects of certain class.}
\usage{
list_objects_of_class(class = NULL, envir = parent.frame(),
  all.names = TRUE)
}
\arguments{
\item{class}{(character|`NULL`) A vector of object classes to return.}

\item{envir}{An environment to look for the objects in.
Other kind of objects will be coerced to an environment.
The default value is parent frame.}

\item{all.names}{(logical)
If `TRUE`, all object names are returned.
If `FALSE`, names which begin with a `.` are omitted.}
}
\value{
A character vector of object names of certain class.
}
\description{
Create a vector of object names of certain class.
}
\keyword{internal}
