% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStudio-addins.R
\name{load_rcmdr_biostat}
\alias{load_rcmdr_biostat}
\title{RStudio addin: Load R Commander in BioStat mode}
\usage{
load_rcmdr_biostat()
}
\description{
Sends library loading commands to the RStudio console.
}
\keyword{internal}
