% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{logicalP}
\alias{logicalP}
\title{Does active dataset contain logicals?}
\usage{
logicalP(n = 1)
}
\arguments{
\item{n}{Minimum number of logical variables}
}
\description{
Return TRUE, if at least n locical variables exist in the active dataset.
}
\keyword{internal}
