% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{make_relative_path}
\alias{make_relative_path}
\title{Make path to relative}
\usage{
make_relative_path(str)
}
\arguments{
\item{str}{(character) Sting (or vector of strings) with absolute path.}
}
\description{
Make absolute path into relative one in respect to current working directory.
}
\keyword{internal}
