% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu-data.R, R/menu-models.R,
%   R/menu-plots.R, R/menu-rows.R, R/menu-settings.R, R/menu-summary.R,
%   R/menu-variables.R, R/online_tools.R, R/rcmdr__active_dataset.R,
%   R/tcltk__refresh_dataset_window_0.R, R/window_data_obj_manage.R,
%   R/window_dataset_bind_cols.R, R/window_dataset_bind_rows.R,
%   R/window_dataset_class.R, R/window_dataset_join.R,
%   R/window_dataset_print_as_md.R, R/window_dataset_select.R,
%   R/window_export_fig_to_pptx.R, R/window_export_to_excel.R,
%   R/window_export_to_rdata.R, R/window_export_to_rds.R,
%   R/window_export_to_text.R, R/window_factor_lvls_drop.R,
%   R/window_fct_relevel.R, R/window_import_from_excel.R,
%   R/window_import_from_pkg.R, R/window_import_from_rds.R,
%   R/window_import_from_text.R, R/window_import_rdata.R, R/window_locale_set.R,
%   R/window_model_select.R, R/window_plots_ggplotly.R, R/window_rows_arrange.R,
%   R/window_rows_col_to_rownames.R, R/window_rows_filter.R,
%   R/window_rows_rm_duplicated.R, R/window_rows_rm_with_na.R,
%   R/window_rows_rowid_to_col.R, R/window_rows_rownames_to_col.R,
%   R/window_rows_slice.R, R/window_summary_count.R, R/window_summary_desc.R,
%   R/window_summary_descr.R, R/window_summary_dfSummary.R,
%   R/window_summary_missings.R, R/window_summary_summary.R,
%   R/window_summary_variables.R, R/window_test_normality.R,
%   R/window_variable_gather.R, R/window_variable_mutate.R,
%   R/window_variable_recode.R, R/window_variable_rename.R,
%   R/window_variable_select.R
\name{Menu-window-functions}
\alias{Menu-window-functions}
\alias{command_dataset_refresh}
\alias{activate_menus}
\alias{command_dataset_view}
\alias{command_dataset_print}
\alias{command_dataset_print_as_df}
\alias{command_dataset_print_as_dt}
\alias{command_dataset_print_as_tibble}
\alias{command_dataset_class}
\alias{command_dataset_as_df}
\alias{command_dataset_as_tibble}
\alias{command_dataset_as_dt}
\alias{command_model_std_lm_coeffs}
\alias{command_model_get_class}
\alias{command_model_print}
\alias{command_model_summary}
\alias{command_model_glance}
\alias{command_model_tidy}
\alias{command_model_augment}
\alias{open_new_plots_window}
\alias{close_all_plots}
\alias{window_show_colors}
\alias{window_plots_image_digitizer}
\alias{command_rownames}
\alias{command_rows_has_rownames}
\alias{command_rows_rm_empty_rows}
\alias{command_getwd}
\alias{command_list_files_wd}
\alias{command_setwd}
\alias{command_openwd}
\alias{command_get_file_info}
\alias{command_list_objects}
\alias{command_session_info_devtools}
\alias{command_session_info_utils}
\alias{window_summary_inspect}
\alias{command_dataset_dim}
\alias{command_dataset_dim_0}
\alias{command_glimpse}
\alias{window_summary_desc_all}
\alias{summary_head_tail}
\alias{summary_var_types}
\alias{summary_var_types_plot}
\alias{command_colnames}
\alias{command_clean_names}
\alias{command_check_names_validity}
\alias{command_all_chr_to_fctr}
\alias{open_online_tool}
\alias{source_online}
\alias{window_online_geogebra_probability}
\alias{window_online_statpages}
\alias{window_online_image_digitizer}
\alias{window_online_mvn}
\alias{window_online_easyROC}
\alias{window_online_r_project}
\alias{window_online_rstudio}
\alias{window_online_bert}
\alias{window_online_gpower}
\alias{window_online_orange}
\alias{window_online_geogebra_download}
\alias{window_online_past_download}
\alias{window_online_tidyverse}
\alias{window_online_broom}
\alias{window_online_dplyr}
\alias{window_online_forcats}
\alias{window_online_magrittr}
\alias{window_online_purrr}
\alias{window_online_readr}
\alias{window_online_readxl}
\alias{window_online_stringr}
\alias{window_online_tidyr}
\alias{window_online_ggplot2}
\alias{window_online_ggstatsplot}
\alias{window_online_ggpubr}
\alias{window_online_plotly}
\alias{window_online_ggplot2_ext}
\alias{window_online_r_chearsheets}
\alias{window_online_sthda}
\alias{window_online_rweekly}
\alias{window_online_r_bloggers}
\alias{window_online_isl}
\alias{window_online_esl}
\alias{window_online_bookdown}
\alias{window_online_biostathandbook}
\alias{window_online_rcompanion}
\alias{window_online_rcompanion_handbook}
\alias{window_online_msmb}
\alias{window_online_r4ds}
\alias{window_online_advanced_r_2ed}
\alias{window_online_rmd}
\alias{window_online_stackoverflow}
\alias{window_online_crossvalidated}
\alias{window_online_homepage}
\alias{window_online_bug_report}
\alias{command_chk_packages_biostat}
\alias{active_dataset}
\alias{refresh_dataset_window_0}
\alias{configure_tags}
\alias{window_data_obj_manage}
\alias{window_dataset_bind_cols}
\alias{window_dataset_bind_rows}
\alias{window_dataset_class}
\alias{window_dataset_join}
\alias{window_dataset_print_as_md}
\alias{window_dataset_print_as_kable}
\alias{window_dataset_select}
\alias{window_export_fig_to_pptx}
\alias{window_export_to_excel}
\alias{window_export_to_rdata}
\alias{window_export_to_rdata_0}
\alias{window_export_to_rds}
\alias{window_export_to_rds_0}
\alias{window_export_to_clipboard}
\alias{window_export_to_text}
\alias{window_factor_lvls_drop}
\alias{window_fct_relevel}
\alias{window_import_from_excel}
\alias{command_list_datasets_in_pkgs}
\alias{window_import_from_pkg}
\alias{window_import_from_rds}
\alias{window_import_from_clipboard}
\alias{window_import_from_text}
\alias{window_import_rdata}
\alias{command_get_locale}
\alias{window_locale_set}
\alias{window_locale_set_0}
\alias{window_model_select}
\alias{window_plots_ggplotly}
\alias{window_rows_arrange}
\alias{window_rows_col_to_rownames0}
\alias{window_rows_col_to_rownames}
\alias{window_rows_filter0}
\alias{window_rows_filter}
\alias{window_rows_rm_duplicated}
\alias{window_rows_rm_with_na}
\alias{window_rows_rowid_to_col}
\alias{window_rows_rownames_to_col}
\alias{window_rows_slice}
\alias{window_summary_count}
\alias{window_summary_desc}
\alias{window_summary_descr}
\alias{window_summary_dfSummary}
\alias{window_summary_missings}
\alias{window_summary_summary}
\alias{window_summary_variables}
\alias{window_test_normality}
\alias{window_variable_gather}
\alias{window_variable_mutate0}
\alias{window_variable_mutate}
\alias{window_variable_recode0}
\alias{window_variable_rename}
\alias{window_variable_select0}
\alias{window_variable_select}
\title{RcmdrPlugin.biostat functions for menus and windows.}
\usage{
command_dataset_refresh()

activate_menus()

command_dataset_view()

command_dataset_print()

command_dataset_print_as_df()

command_dataset_print_as_dt()

command_dataset_print_as_tibble()

command_dataset_class()

command_dataset_as_df()

command_dataset_as_tibble()

command_dataset_as_dt()

command_model_std_lm_coeffs()

command_model_get_class()

command_model_print()

command_model_summary()

command_model_glance()

command_model_tidy()

command_model_augment()

open_new_plots_window()

close_all_plots()

window_show_colors()

window_plots_image_digitizer()

command_rownames()

command_rows_has_rownames()

command_rows_rm_empty_rows()

command_getwd()

command_list_files_wd()

command_setwd()

command_openwd()

command_get_file_info()

command_list_objects()

command_session_info_devtools()

command_session_info_utils()

window_summary_inspect()

command_dataset_dim()

command_dataset_dim_0(.ds)

command_glimpse()

window_summary_desc_all()

summary_head_tail()

summary_var_types()

summary_var_types_plot()

command_colnames()

command_clean_names()

command_check_names_validity()

command_all_chr_to_fctr()

open_online_tool(
  url = NULL,
  copy_to_clipboard = FALSE,
  parent = CommanderWindow()
)

source_online(url = NULL, parent = CommanderWindow())

window_online_geogebra_probability()

window_online_statpages()

window_online_image_digitizer()

window_online_mvn()

window_online_easyROC()

window_online_r_project()

window_online_rstudio()

window_online_bert()

window_online_gpower()

window_online_orange()

window_online_geogebra_download()

window_online_past_download()

window_online_tidyverse()

window_online_broom()

window_online_dplyr()

window_online_forcats()

window_online_magrittr()

window_online_purrr()

window_online_readr()

window_online_readxl()

window_online_stringr()

window_online_tidyr()

window_online_ggplot2()

window_online_ggstatsplot()

window_online_ggpubr()

window_online_plotly()

window_online_ggplot2_ext()

window_online_r_chearsheets()

window_online_sthda()

window_online_rweekly()

window_online_r_bloggers()

window_online_isl()

window_online_esl()

window_online_bookdown()

window_online_biostathandbook()

window_online_rcompanion()

window_online_rcompanion_handbook()

window_online_msmb()

window_online_r4ds()

window_online_advanced_r_2ed()

window_online_rmd()

window_online_stackoverflow()

window_online_crossvalidated()

window_online_homepage()

window_online_bug_report()

command_chk_packages_biostat()

active_dataset(dsname, flushModel = TRUE, flushDialogMemory = TRUE)

refresh_dataset_window_0(
  widget,
  ds_contents,
  preview_type,
  nrow_preview_ds,
  expect_more_rows = FALSE,
  err_msg_default = NULL
)

configure_tags(text_widget)

window_data_obj_manage()

window_dataset_bind_cols()

window_dataset_bind_rows()

window_dataset_class()

window_dataset_join()

window_dataset_print_as_md()

window_dataset_print_as_kable()

window_dataset_select()

window_export_fig_to_pptx()

window_export_to_excel()

window_export_to_rdata()

window_export_to_rdata_0(ds_name = active_dataset())

window_export_to_rds()

window_export_to_rds_0(ds_name = active_dataset())

window_export_to_clipboard()

window_export_to_text()

window_factor_lvls_drop()

window_fct_relevel()

window_import_from_excel()

command_list_datasets_in_pkgs()

window_import_from_pkg()

window_import_from_rds()

window_import_from_clipboard()

window_import_from_text()

window_import_rdata()

command_get_locale()

window_locale_set()

window_locale_set_0(parent = CommanderWindow())

window_model_select()

window_plots_ggplotly()

window_rows_arrange()

window_rows_col_to_rownames0()

window_rows_col_to_rownames(
  new_dsname = NULL,
  init_conditions = NULL,
  incorrect_cond_msg = NULL
)

window_rows_filter0()

window_rows_filter(
  new_dsname = NULL,
  init_conditions = NULL,
  incorrect_cond_msg = NULL
)

window_rows_rm_duplicated()

window_rows_rm_with_na()

window_rows_rowid_to_col()

window_rows_rownames_to_col()

window_rows_slice()

window_summary_count()

window_summary_desc()

window_summary_descr()

window_summary_dfSummary()

window_summary_missings()

window_summary_summary()

window_summary_variables()

window_test_normality()

window_variable_gather()

window_variable_mutate0()

window_variable_mutate()

window_variable_recode0()

window_variable_rename()

window_variable_select0()

window_variable_select(new_dsname = NULL, incorrect_cond_msg = NULL)
}
\description{
Functions that open Rcmdr menus and windows windows.
}
\examples{

library(tidyverse)

df <- tibble::tribble(
  ~x, ~y, ~z, ~q,
  1,   1, 1,  1,
  2,   2,  2, NA,
  3,   3, NA, NA,
  4,  NA, NA, NA,
  NA, NA, NA, NA
)

# remove empty rows
df |> dplyr::filter(if_any(everything(), \(x) !is.na(x)))

# remove rows with all NA in selected vars
df |> dplyr::filter(if_any(c(q, z), \(x) !is.na(x)))

}
\keyword{internal}
