% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-functions.R
\name{Menu-winow-functions}
\alias{Menu-winow-functions}
\alias{dataset_refresh}
\alias{dataset_class}
\alias{dataset_as_df}
\alias{dataset_all_chr_to_fctr}
\alias{dataset_import_rds}
\alias{window_export_as_textfile}
\alias{window_export_as_rds}
\alias{summary_glimpse}
\alias{summary_Hmisc_describe}
\alias{summary_psych_describe}
\title{RcmdrPlugin.BioStat functions for menus and windows}
\usage{
dataset_refresh()

dataset_class()

dataset_as_df()

dataset_all_chr_to_fctr()

dataset_import_rds()

window_export_as_textfile()

window_export_as_rds()

summary_glimpse()

summary_Hmisc_describe()

summary_psych_describe()
}
\description{
Functions that open Rcmdr menus and windows windows
}
\keyword{internal}
