% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{msg_box_check_replace}
\alias{msg_box_check_replace}
\title{Message box to confirm replacement}
\usage{
msg_box_check_replace(name, type = "Variable")
}
\arguments{
\item{name}{string - name of the object to repace.}

\item{type}{string - type of the object to repace.}
}
\value{
Sring eihter "yes" or "no"
}
\description{
Message box to confirm replacement
}
\examples{
msg_box_check_replace()
}
