% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{msg_box_confirm_to_replace}
\alias{msg_box_confirm_to_replace}
\alias{msg_box_confirm_to_replace_all}
\title{Message box to confirm replacement}
\usage{
msg_box_confirm_to_replace(name, type = "Variable", parent = CommanderWindow())

msg_box_confirm_to_replace_all(
  name,
  type = "Variables",
  parent = CommanderWindow()
)
}
\arguments{
\item{name}{string - name of the object to repace.}

\item{type}{string - type of the object to repace.}

\item{parent}{Parent Tcl/Tk frame..}
}
\value{
Sring eihter "yes" or "no"
}
\description{
Message box to confirm replacement
}
\examples{
\dontrun{
msg_box_confirm_to_replace()
}
}
