% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_esquisse_app.R
\name{open_esquisse_app}
\alias{open_esquisse_app}
\title{Open the esquisse interactive ggplot builder}
\usage{
open_esquisse_app()
}
\description{
Launches esquisse with the active dataset in the RStudio viewer.
}
\keyword{internal}
