% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_aaa-package.R
\docType{package}
\name{RcmdrPlugin.BioStat-package}
\alias{RcmdrPlugin.BioStat-package}
\alias{RcmdrPlugin.BioStat}
\alias{BioStat}
\title{Rcmdr Plug-In Package}
\description{
This package is an R Commander plug-in package in GUI interface.
}
\seealso{
\code{\link[Rcmdr:Rcmdr]{Rcmdr}},
\code{\link[ggplot2:ggplot]{ggplot2}},
\code{\link[survival:survfit]{survfit}},
\code{\link[RColorBrewer:RColorBrewer]{RColorBrewer}}
\code{\link[ggthemes:ggthemes]{ggthemes}}
}
\keyword{hplot}
\keyword{package}
