% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_all_fct.R
\name{window_all_fct}
\alias{window_all_fct}
\title{Wrapper Function of Factorize Subclass}
\usage{
window_all_fct()
}
\description{
\code{window_all_fct} function is a wrapper function of \code{factorize} class for the R-commander menu bar.
}
\keyword{hplot}
