% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_plot-factorise.R
\name{windowFactorize}
\alias{windowFactorize}
\title{Wrapper Function of Factorize Subclass}
\usage{
windowFactorize()
}
\description{
\code{windowFactorize} function is a wrapper function of \code{factorize} class for the R-commander menu bar.
}
\keyword{hplot}
