% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gbox.R
\name{windowBox}
\alias{windowBox}
\title{Wrapper Function of Box Plot Subclass}
\usage{
windowBox()
}
\description{
\code{windowBox} function is a wrapper function of \code{gbox} class for the R-commander menu bar.
}
\keyword{hplot}

