% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gcont.R
\name{windowContour}
\alias{windowContour}
\title{Wrapper Function of Contour Plot Subclass}
\usage{
windowContour()
}
\description{
\code{windowContour} function is a wrapper function of \code{gcont} class for the R-commander menu bar.
}
\keyword{hplot}
