% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gDensity.R
\name{windowDensity}
\alias{windowDensity}
\title{Wrapper Function of Density plot Subclass}
\usage{
windowDensity()
}
\description{
\code{windowDensity} function is a wrapper function of \code{gDensity} class for the R-commander menu bar.
}
\keyword{hplot}

