% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdiscbar.r
\name{windowDiscretebar}
\alias{windowDiscretebar}
\title{Wrapper Function of Discrete Bar Charts Subclass}
\usage{
windowDiscretebar()
}
\description{
\code{windowDiscretebar} function is a wrapper function of \code{gdiscbar} class for the R-commander menu bar.
}
\keyword{hplot}

