% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistBeta}
\alias{windowDistBeta}
\title{Wrapper Function of Beta Distribution Plot Subclass}
\usage{
windowDistBeta()
}
\description{
\code{windowDistBeta} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
