% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistBinom}
\alias{windowDistBinom}
\title{Wrapper Function of Binomial Distribution Plot Subclass}
\usage{
windowDistBinom()
}
\description{
\code{windowDistBinom} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
