% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistCauchy}
\alias{windowDistCauchy}
\title{Wrapper Function of Cauchy Distribution Plot Subclass}
\usage{
windowDistCauchy()
}
\description{
\code{windowDistCauchy} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
