% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistChisq}
\alias{windowDistChisq}
\title{Wrapper Function of Chi-square Distribution Plot Subclass}
\usage{
windowDistChisq()
}
\description{
\code{windowDistChisq} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
