% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistExp}
\alias{windowDistExp}
\title{Wrapper Function of Exponential Distribution Plot Subclass}
\usage{
windowDistExp()
}
\description{
\code{windowDistExp} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
