% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistF}
\alias{windowDistF}
\title{Wrapper Function of F Distribution Plot Subclass}
\usage{
windowDistF()
}
\description{
\code{windowDistF} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

