% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.r
\name{windowDistGamma}
\alias{windowDistGamma}
\title{Wrapper Function of Gamma Distribution Plot Subclass}
\usage{
windowDistGamma()
}
\description{
\code{windowDistGamma} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

