% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistGeom}
\alias{windowDistGeom}
\title{Wrapper Function of Geometric Distribution Plot Subclass}
\usage{
windowDistGeom()
}
\description{
\code{windowDistGeom} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
