% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.r
\name{windowDistHyper}
\alias{windowDistHyper}
\title{Wrapper Function of Hypergeometric Distribution Plot Subclass}
\usage{
windowDistHyper()
}
\description{
\code{windowDistHyper} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

