% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistLnorm}
\alias{windowDistLnorm}
\title{Wrapper Function of Log-normal Distribution Plot Subclass}
\usage{
windowDistLnorm()
}
\description{
\code{windowDistLnorm} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
