% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.r
\name{windowDistLogis}
\alias{windowDistLogis}
\title{Wrapper Function of Logistic Distribution Plot Subclass}
\usage{
windowDistLogis()
}
\description{
\code{windowDistLogis} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

