% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.r
\name{windowDistNbinom}
\alias{windowDistNbinom}
\title{Wrapper Function of Negative Binomial Distribution Plot Subclass}
\usage{
windowDistNbinom()
}
\description{
\code{windowDistNbinom} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

