% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.r
\name{windowDistNorm}
\alias{windowDistNorm}
\title{Wrapper Function of Normal Distribution Plot Subclass}
\usage{
windowDistNorm()
}
\description{
\code{windowDistNorm} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

