% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistPois}
\alias{windowDistPois}
\title{Wrapper Function of Poisson Distribution Plot Subclass}
\usage{
windowDistPois()
}
\description{
\code{windowDistPois} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

