% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.r
\name{windowDistT}
\alias{windowDistT}
\title{Wrapper Function of t Distribution Plot Subclass}
\usage{
windowDistT()
}
\description{
\code{windowDistT} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

