% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.R
\name{windowDistUnif}
\alias{windowDistUnif}
\title{Wrapper Function of Uniform Distribution Plot Subclass}
\usage{
windowDistUnif()
}
\description{
\code{windowDistUnif} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}
