% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gdist.r
\name{windowDistWeibull}
\alias{windowDistWeibull}
\title{Wrapper Function of Weibull Distribution Plot Subclass}
\usage{
windowDistWeibull()
}
\description{
\code{windowDistWeibull} function is a wrapper function of \code{gdist} class for the R-commander menu bar.
}
\keyword{hplot}

