% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ghist.r
\name{windowHist}
\alias{windowHist}
\title{Wrapper Function of Histogram Subclass}
\usage{
windowHist()
}
\description{
\code{windowHist} function is a wrapper function of \code{ghist} class for the R-commander menu bar.
}
\keyword{hplot}

