% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gkm.R
\name{windowKM}
\alias{windowKM}
\title{Wrapper Function of Kaplan-Meier Plot Subclass}
\usage{
windowKM()
}
\description{
\code{windowKM} function is a wrapper function of \code{gkm} class for the R-commander menu bar.
}
\keyword{hplot}
