% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gline.R
\name{windowLine}
\alias{windowLine}
\title{Wrapper Function of Line Chart Subclass}
\usage{
windowLine()
}
\description{
\code{windowScatter} function is a wrapper function of \code{gline} class for the R-commander menu bar.
}
\keyword{hplot}

