% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gqq.R
\name{windowQQ}
\alias{windowQQ}
\title{Wrapper Function of Q-Q Plot Subclass}
\usage{
windowQQ()
}
\description{
\code{windowQQ} function is a wrapper function of \code{gqq} class for the R-commander menu bar.
}
\keyword{hplot}
