% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gscat.R
\name{windowScatter}
\alias{windowScatter}
\title{Wrapper Function of Scatter Plot Subclass}
\usage{
windowScatter()
}
\description{
\code{windowScatter} function is a wrapper function of \code{gscat} class for the R-commander menu bar.
}
\keyword{hplot}
