% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gscatmat.R
\name{windowScattermat}
\alias{windowScattermat}
\title{Wrapper Function of Scatter Plot Matrix Subclass}
\usage{
windowScattermat()
}
\description{
\code{windowScattermat} function is a wrapper function of \code{gscatmat} class for the R-commander menu bar.
}
\keyword{hplot}
