% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aaa.R
\docType{methods}
\name{getXlab,plot_base-method}
\alias{getXlab,plot_base-method}
\title{Get \code{xlab}}
\usage{
\S4method{getXlab}{plot_base}(parms)
}
\arguments{
\item{parms}{List of objects; graphics settings parameters.}
}
\value{
\describe{
\item{xlab}{Character; a code snippet for the x-axis label.}
}
}
\description{
\code{getXlab} method gets the x-axis label.
}
\seealso{
Other plot: \code{\link{checkError,plot_base-method}},
  \code{\link{checkTheme,plot_base-method}},
  \code{\link{checkVariable,plot_base-method}},
  \code{\link{commandDoIt,plot_base-method}},
  \code{\link{factorize-class}},
  \code{\link{gDensity-class}}, \code{\link{gbox-class}},
  \code{\link{gcont-class}}, \code{\link{gdiscbar-class}},
  \code{\link{gdist-class}},
  \code{\link{getCoord,plot_base-method}},
  \code{\link{getFacet,plot_base-method}},
  \code{\link{getGeom,plot_base-method}},
  \code{\link{getGgplot,plot_base-method}},
  \code{\link{getHelp,plot_base-method}},
  \code{\link{getMain,plot_base-method}},
  \code{\link{getMessage,plot_base-method}},
  \code{\link{getOpts,plot_base-method}},
  \code{\link{getParms,plot_base-method}},
  \code{\link{getPlot,plot_base-method}},
  \code{\link{getScale,plot_base-method}},
  \code{\link{getTheme,plot_base-method}},
  \code{\link{getWindowTitle,plot_base-method}},
  \code{\link{getYlab,plot_base-method}},
  \code{\link{getZlab,plot_base-method}},
  \code{\link{ghist-class}}, \code{\link{gkm-class}},
  \code{\link{gline-class}}, \code{\link{gpie-class}},
  \code{\link{gqq-class}}, \code{\link{gscat-class}},
  \code{\link{gscatmat-class}},
  \code{\link{plotWindow,plot_base-method}},
  \code{\link{plot_base-class}},
  \code{\link{registRmlist,plot_base-method}},
  \code{\link{removeRmlist,plot_base-method}},
  \code{\link{savePlot,plot_base-method}},
  \code{\link{setBack,plot_base-method}},
  \code{\link{setDataframe,plot_base-method}},
  \code{\link{setFront,plot_base-method}}
}
\keyword{hplot}
