% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_plot-aaa.R
\docType{methods}
\name{removeRmlist,plot_base-method}
\alias{removeRmlist,plot_base-method}
\title{Remove Registered Temporary Objects}
\usage{
\S4method{removeRmlist}{plot_base}()
}
\description{
\code{removeRmlist} method removes registered temporary objects.
}
\seealso{
Other plot: \code{\link{checkError,plot_base-method}},
  \code{\link{checkTheme,plot_base-method}},
  \code{\link{checkVariable,plot_base-method}},
  \code{\link{commandDoIt,plot_base-method}},
  \code{\link{factorize-class}},
  \code{\link{getCoord,plot_base-method}},
  \code{\link{getFacet,plot_base-method}},
  \code{\link{getGeom,plot_base-method}},
  \code{\link{getGgplot,plot_base-method}},
  \code{\link{getHelp,plot_base-method}},
  \code{\link{getMain,plot_base-method}},
  \code{\link{getMessage,plot_base-method}},
  \code{\link{getOpts,plot_base-method}},
  \code{\link{getParms,plot_base-method}},
  \code{\link{getPlot,plot_base-method}},
  \code{\link{getScale,plot_base-method}},
  \code{\link{getTheme,plot_base-method}},
  \code{\link{getWindowTitle,plot_base-method}},
  \code{\link{getXlab,plot_base-method}},
  \code{\link{getYlab,plot_base-method}},
  \code{\link{getZlab,plot_base-method}},
  \code{\link{ghist-class}},
  \code{\link{plotWindow,plot_base-method}},
  \code{\link{plot_base-class}},
  \code{\link{registRmlist,plot_base-method}},
  \code{\link{savePlot,plot_base-method}},
  \code{\link{setBack,plot_base-method}},
  \code{\link{setDataframe,plot_base-method}},
  \code{\link{setFront,plot_base-method}}
}
\keyword{hplot}
