% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_plot-aaa.R
\docType{class}
\name{plot_base-class}
\alias{plot_base-class}
\alias{plot_base}
\title{Base Class for Plot Windows}
\description{
\code{plot_base} class is the base class for plot windows.
}
\details{
This class is a framework for implementing subclasses which show dialog boxes for graphics editing.
}

\section{Fields}{

\describe{
\item{\code{top}: }{\code{tkwin} class object; parent of widget window.}
\item{\code{alternateFrame}: }{\code{tkwin} class object; a special frame for some GUI parts.}
\item{\code{rmlist}: }{List of character; deletable temporary objects.}
\item{\code{mode}: }{numeric; the executive mode (0 = justDoIt, 1 = doItAndPrint).}
}
}

\section{Contains}{

NULL
}

\section{Methods}{

\describe{
\item{\code{plotWindow()}: }{Create the window that make plots.}
\item{\code{savePlot(plot)}: }{Save the plot.}
\item{\code{registRmlist(object)}: }{Register deletable temporary objects.}
\item{\code{removeRmlist()}: }{Remove registered temporary objects.}
\item{\code{setFront()}: }{Set front parts of frames.}
\item{\code{setBack()}: }{Set back parts of frames.}
\item{\code{getWindowTitle()}: }{Get the title of the window.}
\item{\code{getHelp()}: }{Get the title of the help document.}
\item{\code{getParms()}: }{Get graphics settings parameters.}
\item{\code{checkTheme(index)}: }{Check themes.}
\item{\code{checkVariable(var)}: }{Check a variable length.}
\item{\code{checkError(parms)}: }{Check errors.}
\item{\code{setDataframe(parms)}: }{Set data frames.}
\item{\code{getGgplot(parms)}: }{Get \code{ggplot}.}
\item{\code{getGeom(parms)}: }{Get \code{geom}.}
\item{\code{getScale(parms)}: }{Get \code{scale}.}
\item{\code{getCoord(parms)}: }{Get \code{coord}.}
\item{\code{getFacet(parms)}: }{Get \code{facet}.}
\item{\code{getXlab(parms)}: }{Get \code{xlab}.}
\item{\code{getYlab(parms)}: }{Get \code{ylab}.}
\item{\code{getZlab(parms)}: }{Get \code{zlab}.}
\item{\code{getMain(parms)}: }{Get the main label.}
\item{\code{getTheme(parms)}: }{Get \code{theme}.}
\item{\code{getOpts(parms)}: }{Get other \code{opts}.}
\item{\code{getPlot(parms)}: }{Get the plot object.}
\item{\code{getMessage()}: }{Get the plot error message.}
\item{\code{commandDoIt(command)}: }{An wrapper function for command execution.}
}
}

\seealso{
Other plot: \code{\link{checkError,plot_base-method}},
  \code{\link{checkTheme,plot_base-method}},
  \code{\link{checkVariable,plot_base-method}},
  \code{\link{commandDoIt,plot_base-method}},
  \code{\link{factorize-class}},
  \code{\link{getCoord,plot_base-method}},
  \code{\link{getFacet,plot_base-method}},
  \code{\link{getGeom,plot_base-method}},
  \code{\link{getGgplot,plot_base-method}},
  \code{\link{getHelp,plot_base-method}},
  \code{\link{getMain,plot_base-method}},
  \code{\link{getMessage,plot_base-method}},
  \code{\link{getOpts,plot_base-method}},
  \code{\link{getParms,plot_base-method}},
  \code{\link{getPlot,plot_base-method}},
  \code{\link{getScale,plot_base-method}},
  \code{\link{getTheme,plot_base-method}},
  \code{\link{getWindowTitle,plot_base-method}},
  \code{\link{getXlab,plot_base-method}},
  \code{\link{getYlab,plot_base-method}},
  \code{\link{getZlab,plot_base-method}},
  \code{\link{ghist-class}},
  \code{\link{plotWindow,plot_base-method}},
  \code{\link{registRmlist,plot_base-method}},
  \code{\link{removeRmlist,plot_base-method}},
  \code{\link{savePlot,plot_base-method}},
  \code{\link{setBack,plot_base-method}},
  \code{\link{setDataframe,plot_base-method}},
  \code{\link{setFront,plot_base-method}}
}
\keyword{An}
\keyword{Back}
\keyword{Check}
\keyword{Coord}
\keyword{Error}
\keyword{Facet}
\keyword{Front}
\keyword{Geom}
\keyword{Get}
\keyword{Ggplot}
\keyword{Help}
\keyword{List}
\keyword{Main}
\keyword{Message}
\keyword{Opts}
\keyword{Parameters}
\keyword{Plot}
\keyword{Register}
\keyword{Remove}
\keyword{Save}
\keyword{Scale}
\keyword{Set}
\keyword{Theme}
\keyword{Title}
\keyword{Window}
\keyword{Windows}
\keyword{Xlab}
\keyword{Ylab}
\keyword{Zlab}
\keyword{\code{data.frame}}
\keyword{\code{rm()}}
\keyword{command}
\keyword{execution}
\keyword{for}
\keyword{function}
\keyword{hplot}
\keyword{length.}
\keyword{themes.}
\keyword{variable}
\keyword{wrapper}
