% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package.R
\docType{package}
\name{RcmdrPlugin.biostat}
\alias{RcmdrPlugin.biostat-package}
\alias{RcmdrPlugin.biostat}
\title{R Commander Plug-in for Basic Data Management and Analysis Tasks}
\description{
\pkg{RcmdrPlugin.biostat} is an
\href{\pkg{Rcmdr}}{https://CRAN.R-project.org/package=Rcmdr}
plug-in for the most common data wrangling, visualization and analysis tasks
using "tidyverse" family functions as well as functions from other packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gegznav.github.io/RcmdrPlugin.biostat/}
  \item \url{https://github.com/GegznaV/RcmdrPlugin.biostat/}
  \item Report bugs at \url{https://github.com/GegznaV/RcmdrPlugin.biostat/issues}
}

}
\author{
\strong{Maintainer}: Vilmantas Gegzna \email{GegznaV@gmail.com} (\href{https://orcid.org/0000-0002-9500-5167}{ORCID})

Other contributors:
\itemize{
  \item Yoshinobu Kanda [contributor]
  \item John Fox [contributor]
  \item Milan Bouchet-Valat [contributor]
  \item Liviu Andronic [contributor]
  \item Michael Ash [contributor]
  \item Theophilius Boye [contributor]
  \item Stefano Calza [contributor]
  \item Andy Chang [contributor]
  \item Philippe Grosjean [contributor]
  \item Richard Heiberger [contributor]
  \item Kosar Karimi Pour [contributor]
  \item G. Jay Kerns [contributor]
  \item Renaud Lancelot [contributor]
  \item Matthieu Lesnoff [contributor]
  \item Uwe Ligges [contributor]
  \item Samir Messad [contributor]
  \item Martin Maechler [contributor]
  \item Robert Muenchen [contributor]
  \item Duncan Murdoch [contributor]
  \item Erich Neuwirth [contributor]
  \item Dan Putler [contributor]
  \item Brian Ripley [contributor]
  \item Miroslav Ristic [contributor]
  \item Peter Wolf [contributor]
  \item Kevin Wright [contributor]
}

}
\keyword{internal}
