% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{scale_vector}
\alias{scale_vector}
\title{Scale a vector}
\usage{
scale_vector(x, center = mean, scale = sd)
}
\arguments{
\item{x}{A numeric vector.}

\item{center}{Eiter a function that computes center of data
(such as \code{mean}) or a simgle numeric value.}

\item{scale}{Eiter a function that computes variability of data
(such as \code{sd}) or a simgle numeric value.}
}
\value{
A scaled vector
}
\description{
Subtract center and divide by scale.
}
\examples{

x = 1:10
scale_vector(x)

scale_vector(x, center = median, scale = IQR)

scale_vector(x, center = 10, scale = 2)

}
