% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gDensity.R
\name{tab}
\alias{tab}
\title{Instert tab symbol}
\usage{
tab(n = 1)
}
\arguments{
\item{n}{(integer) number of tabs to be inserted}
}
\value{
A string.
}
\description{
Instert tab symbol
}
\examples{

tab()
# "\\t"

tab(3)
# "\\t\\t\\t"

}

