% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_elements.R
\name{tcl_get_property}
\alias{tcl_get_property}
\title{Get Vaalue of Tcl/Tk Widget Property}
\usage{
tcl_get_property(.widget, property)
}
\arguments{
\item{.widget}{Tcl/Tk widget.}

\item{property}{(character) name of a property (e.g., "-text", "-image").}
}
\description{
Get Vaalue of Tcl/Tk Widget Property
}
\keyword{internal}
