% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_elements.R
\name{TclTk-helper-functions}
\alias{TclTk-helper-functions}
\alias{labeled_frame}
\alias{tk_title}
\alias{tcl_get_parent}
\alias{tcl_get_children_id}
\alias{tcl_get_siblings_id}
\alias{tcl_get_obj_by_id}
\title{Helper functions for Tcl/Tk.}
\usage{
labeled_frame(parent, label = NULL, ...)

tk_title(parent = top, text = "", pady = c(5, 9),
  font = tkfont.create(weight = "bold", size = 9),
  fg = Rcmdr::getRcmdr("title.color"), ...)

tcl_get_parent(obj)

tcl_get_children_id(obj)

tcl_get_siblings_id(obj)

tcl_get_obj_by_id(id, main_win = CommanderWindow())
}
\description{
Helper functions for Tcl/Tk.
}
\keyword{internal}
