% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_elements.R
\name{TclTk-labels}
\alias{TclTk-labels}
\alias{tk_label}
\alias{tk_label_blue}
\alias{tk_label_red}
\title{Tcl/Tk labels}
\usage{
tk_label(parent, text = "", ..., fg = NULL)

tk_label_blue(parent, text = "", ..., fg = Rcmdr::getRcmdr("title.color"))

tk_label_red(parent, text = "", ..., size = 8, weight = "bold", fg = "darkred")
}
\arguments{
\item{parent}{(\code{"tkwin"} object) Parent Tcl/Tk window or frame.}

\item{text}{(character) Label text.}

\item{...}{Other arguments to pass to \code{tcltk2::tk2label()}.}

\item{fg}{(character) Foreground color.}

\item{size}{(integer) Font size}

\item{weight}{(character) Font weight.}
}
\description{
Functions, that create Tcl/Tk labels.
\itemize{
\item \code{tk_label()} Uses the default color for labels (usually, black).
\item \code{tk_label_blue()} Uses the default color for titles (usually, blue).
}
}
\seealso{
\code{tcltk2::tk2label()}, \code{Rcmdr::labelRcmdr()}, \code{tcltk::ttklabel()}.
}
