% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-natrisk21.R
\name{theme_natrisk21}
\alias{theme_natrisk21}
\title{A Theme for No of At Risk in Kaplan-Meier Plots}
\usage{
theme_natrisk21(base_theme, base_size = 20, base_family = "")
}
\arguments{
\item{base_theme}{base theme name}

\item{base_size}{base font size}

\item{base_family}{font family name}
}
\description{
These functions generate graphics themes for no of at risk in Kaplan-Meier Plots.
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{color}

