% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-wsj2.r
\name{theme_wsj2}
\alias{theme_wsj2}
\title{A Wrapper Theme for ggthemes's theme_wsj}
\usage{
theme_wsj2(base_size = 16, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{font family name}
}
\description{
A Wrapper Theme for ggthemes's theme_wsj
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot}} \code{\link[ggthemes:theme_wsj]{theme_wsj}}
}
\keyword{color}

