% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_commands.R
\name{tk_activate}
\alias{tk_activate}
\alias{tk_get_state.default}
\alias{tk_activate.default}
\alias{tk_disable.default}
\alias{tk_normalize.default}
\title{Activate/Disable TK Objects}
\usage{
\method{tk_get_state}{default}(tk_obj, ...)

\method{tk_activate}{default}(tk_obj, ...)

\method{tk_disable}{default}(tk_obj, ...)

\method{tk_normalize}{default}(tk_obj, ...)
}
\arguments{
\item{tk_obj}{TK object.}

\item{...}{other options to be passed to \code{tkconfigure}.}
}
\description{
Modify state of tk objects.
}
\examples{
# tk_activate(tk_obj)
# tk_disable(tk_obj)

}
\keyword{internal}
