% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk__tktag.R
\name{tktag_add}
\alias{tktag_add}
\alias{tktag_add_first}
\alias{tktag_add_row}
\title{Add tags to Tk text widgets.}
\usage{
tktag_add(obj, tag, pattern, row_ind = NULL, all = TRUE)

tktag_add_first(obj, pattern, tag, row_ind = NULL)

tktag_add_row(obj, pattern, tag)
}
\arguments{
\item{obj}{Tk text widget (Tk object).}

\item{tag}{(string) The namo of a tag to add.}

\item{pattern}{(string) Text pattern to add text at.}

\item{row_ind}{(integer) Row indices.}

\item{all}{(logical) If \code{TRUE}, adds tag to all matches, if \code{FALSE}, adds tag
to the first match only.}
}
\description{
Add TK text tags to Tk text widgets:
\code{tktag_add()} adds tags to the places of text that match a pattern.
\code{tktag_add_first()} \emph{in each row}, adds tags to the \emph{first} occurance that
matches a pattern.
\code{tktag_add_row()} adds text tags to the \strong{rows} that match a pattern.
}
\seealso{
\code{\link[tcltk:TkWidgetcmds]{tcltk::tktag.add()}}
}
