% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-gettextKmg2.R
\name{gettext_Bio}
\alias{gettext_Bio}
\title{A \code{gettextRcmdr} Wrapper Function}
\usage{
gettext_Bio(...)
}
\arguments{
\item{...}{arguments passed to gettext function}
}
\description{
This function is a \code{gettextRcmdr} wrapper function for this package.
}
\seealso{
\code{\link[Rcmdr:Rcmdr.Utilities]{Rcmdr.Utilities}}
}
\keyword{documentation}
