% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_util-nonFactors.R
\name{nonFactors}
\alias{nonFactors}
\title{Select Non Factor Variables}
\usage{
nonFactors()
}
\description{
This is a function to select non-factor variables for the \code{Rcmdr} package.
}
\seealso{
\code{\link[Rcmdr:Rcmdr.Utilities]{Rcmdr.Utilities}}
}
\keyword{hplot}
