% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_util-nonFactors.R
\name{nonFactorsP}
\alias{nonFactorsP}
\title{Check Non Factor Variables in The Active Dataset}
\usage{
nonFactorsP(n = 1)
}
\arguments{
\item{n}{Numeric: the target number of \code{n}.}
}
\value{
\describe{
\item{result}{Boolean; the active dataset has or does not have more than or equal to \code{n} non-factor variables.}
}
}
\description{
This is a function to check the active dataset has more than or equal to \code{n} non-factor variables.
}
\seealso{
\code{\link[Rcmdr:Rcmdr.Utilities]{Rcmdr.Utilities}}
}
\keyword{hplot}
