% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{var_pos_n}
\alias{var_pos_n}
\title{...}
\usage{
var_pos_n(variables, type = c("all", "numeric", "factor", "factor_strict",
  "twoLevelFactor", "nonfactor", "character", "logical"), vars = NULL)
}
\description{
...
}
\keyword{internal}
