% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{variables_fct2_like}
\alias{variables_fct2_like}
\title{Two-level factor names in active dataset}
\usage{
variables_fct2_like()
}
\description{
Two-level factor names in active dataset
}
\keyword{internal}
