% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{variables_fct_like_2_lvls}
\alias{variables_fct_like_2_lvls}
\title{Two-level factor names in active dataset}
\usage{
variables_fct_like_2_lvls()
}
\description{
Two-level factor names in active dataset
}
\keyword{internal}
