% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{variables_nfct}
\alias{variables_nfct}
\title{Non-factor-like variable names in active dataset}
\usage{
variables_nfct()
}
\description{
Non-factor-like variable names in active dataset
}
\keyword{internal}
