% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{variables_non_fct}
\alias{variables_non_fct}
\title{Non-factor variable names in active dataset}
\usage{
variables_non_fct()
}
\description{
Non-factor variable names in active dataset
}
\keyword{internal}
