% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{variables_non_fct_like}
\alias{variables_non_fct_like}
\title{Non-factor-like variable names in active dataset}
\usage{
variables_non_fct_like()
}
\description{
Non-factor-like variable names in active dataset
}
\keyword{internal}
