% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{variablesP}
\alias{variablesP}
\title{Does dataset contain certain number of variables?}
\usage{
variablesP(n = 1)
}
\arguments{
\item{n}{Minimum number of character variables}
}
\description{
Return TRUE, if at least n variables exist in the active dataset.
}
\keyword{internal}
