% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R, R/tcltk_commands.R
\name{widget-state}
\alias{widget-state}
\alias{tk_get_default_enabled_state}
\alias{tk_set_default_enabled_state}
\alias{tk_get_state}
\alias{tk_set_state}
\alias{tk_normalize}
\alias{tk_activate}
\alias{tk_read_only}
\alias{tk_enable}
\alias{tk_disable}
\alias{tk_get_default_enabled_state.default}
\alias{tk_get_state.default}
\alias{tk_set_default_enabled_state.default}
\alias{tk_set_state.default}
\alias{tk_normalize.default}
\alias{tk_activate.default}
\alias{tk_read_only.default}
\alias{tk_disable.default}
\title{Get and set state of Tk widgets.}
\usage{
tk_get_default_enabled_state(obj, ...)

tk_set_default_enabled_state(obj, state, ...)

tk_get_state(obj, ...)

tk_set_state(obj, state, ...)

tk_normalize(obj, ...)

tk_activate(obj, ...)

tk_read_only(obj, ...)

tk_enable(obj, ...)

tk_disable(obj, ...)

\method{tk_get_default_enabled_state}{default}(obj)

\method{tk_get_state}{default}(obj, ...)

\method{tk_set_default_enabled_state}{default}(obj, state = c("active", "normal", "readonly"))

\method{tk_set_state}{default}(obj, state, ...)

\method{tk_normalize}{default}(obj, ...)

\method{tk_activate}{default}(obj, ...)

\method{tk_read_only}{default}(obj, ...)

\method{tk_disable}{default}(obj, ...)
}
\arguments{
\item{obj}{Tk widget or a string with the ID of an existing Tk widget
(e.g., \code{".1.2.5"}).}

\item{...}{other options to be passed to \code{\link[tcltk:tkconfigure]{tcltk::tkconfigure()}}.}

\item{state}{(character) The state of widget.
Usually one of "normal", "active", "diabled", "readonly".}
}
\description{
Get or modify the state of Tk widgets.
}
\details{
The function to get widget's state:
\itemize{
\item \code{tk_get_state()}
}

The functions to set widget's state to:
\itemize{
\item \code{tk_normalize()} to "normal";
\item \code{tk_activate()} to "active";
\item \code{tk_read_only()} to "readonly";
\item \code{tk_disable()} to "disable";
\item \code{tk_enable()} to default non-disabled state;
\item \code{tk_set_state()} to the indicated state.
}

The functions to manage widget's state to:
\itemize{
\item \code{tk_get_default_enabled_state()} get default default enabled state. If not
set, "normal" is returned.
\item \code{tk_set_default_enabled_state()} change the default enabled state to
either "normal", "active" or "readonly".
}
}
\examples{
# tk_get_state(obj)
# tk_get_state(".1.2.5")

# tk_disable(obj)
# tk_normalize(obj)
# tk_set_state(obj, "normal")
}
\seealso{
\code{\link[tcltk:tkconfigure]{tcltk::tkconfigure()}}
}
\keyword{internal}
