% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_about_package.R
\name{window_about_biostat_version}
\alias{window_about_biostat_version}
\title{window_about_biostat_version}
\usage{
window_about_biostat_version()
}
\description{
window_about_biostat_version
}
\keyword{internal}
