% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdr__bin_variable.R
\name{window_bin_variable}
\alias{window_bin_variable}
\title{Bin a numeric variable into intervals (Rcmdr-based)}
\usage{
window_bin_variable()
}
\description{
Based on Rcmdr function `binVariable()`. Authors: Dan Putler, J. Fox.
}
\keyword{internal}
