% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_convert_any.R
\name{window_convert_any}
\alias{window_convert_any}
\title{Rcmdr windows for variable class conversion}
\usage{
window_convert_any()
}
\description{
Rcmdr windows for variable class conversion
}
\seealso{
Other conversion: \code{\link{window_variable_convert_type}}
}
\concept{conversion}
\keyword{internal}
