% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_dataset_copy.R
\name{window_dataset_copy}
\alias{window_dataset_copy}
\title{command_new_dataset}
\usage{
window_dataset_copy()
}
\description{
Code of this function is taken from `Rcmdr` package and slightly modified.
}
\keyword{internal}
