% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdr__onEdit.R
\name{window_dataset_edit_rcmdr}
\alias{window_dataset_edit_rcmdr}
\title{Edit the active dataset (Rcmdr-based)}
\usage{
window_dataset_edit_rcmdr()
}
\description{
Based on function `onEdit` from Rcmdr 2.5-1.
}
\keyword{internal}
