% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdr__window_dataset_new.R
\name{window_dataset_new_rcmdr}
\alias{window_dataset_new_rcmdr}
\title{command_new_dataset}
\usage{
window_dataset_new_rcmdr()
}
\description{
Code of this function is taken from `Rcmdr` package and slightly modified.
}
\keyword{internal}
