% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_dataset_rename.R
\name{window_dataset_rename}
\alias{window_dataset_rename}
\title{command_new_dataset}
\usage{
window_dataset_rename()
}
\description{
Code of this function is taken from `Rcmdr` package and slightly modified.
}
\keyword{internal}
