% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdr__onView.R
\name{window_dataset_view_rcmdr}
\alias{window_dataset_view_rcmdr}
\title{View the active dataset (Rcmdr-based)}
\usage{
window_dataset_view_rcmdr()
}
\description{
Based on function `onView` from Rcmdr 2.5-1.
}
\keyword{internal}
