% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_ggThemeAssist.R
\name{window_ggThemeAssist}
\alias{window_ggThemeAssist}
\title{window_ggThemeAssist}
\usage{
window_ggThemeAssist()
}
\description{
This function creates an example code for `ggThemeAssist`.
}
\keyword{internal}
