% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_lgl_convert.R
\name{window_lgl_convert}
\alias{window_lgl_convert}
\title{Rcmdr windows for variable class conversion}
\usage{
window_lgl_convert()
}
\description{
Rcmdr windows for variable class conversion
}
\seealso{
Other conversion: \code{\link{window_chr_convert}},
  \code{\link{window_convert_all}},
  \code{\link{window_convert_any}},
  \code{\link{window_fct_convert}},
  \code{\link{window_num_convert}}
}
\concept{conversion}
\keyword{internal}
