% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_log_transform.R
\name{window_log_transform}
\alias{window_log_transform}
\title{Rcmdr window for log transformation}
\usage{
window_log_transform()
}
\description{
Rcmdr window for log transformation
}
\seealso{
Other transformations: \code{\link{window_z_transform}}
}
\keyword{internal}
