% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_new_dataset.R
\name{window_new_dataset_rcmdr}
\alias{window_new_dataset_rcmdr}
\title{command_new_dataset}
\usage{
window_new_dataset_rcmdr()
}
\description{
This function is taken from `Rcmdr` package and slightly modified.
}
\keyword{internal}
