% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_num_convert--new.R,
%   R/window_num_convert.R
\name{window_num_convert}
\alias{window_num_convert}
\title{Rcmdr windows for variable class conversion}
\usage{
window_num_convert()

window_num_convert()
}
\description{
Rcmdr windows for variable class conversion

Rcmdr windows for variable class conversion
}
\seealso{
Other conversion: \code{\link{window_chr_convert}},
  \code{\link{window_fct_convert}},
  \code{\link{window_lgl_convert}}

Other conversion: \code{\link{window_chr_convert}},
  \code{\link{window_fct_convert}},
  \code{\link{window_lgl_convert}}
}
\concept{conversion}
\keyword{internal}
