% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_num_transform_log.R
\name{window_num_transform_log}
\alias{window_num_transform_log}
\title{Rcmdr window for log transformation}
\usage{
window_num_transform_log()
}
\description{
Rcmdr window for log transformation
}
\seealso{
Other transformations: 
\code{\link{window_num_transform_z}()}
}
\concept{transformations}
\keyword{internal}
