% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_num_transform_z.R
\name{window_num_transform_z}
\alias{window_num_transform_z}
\title{Rcmdr window for z transformation}
\usage{
window_num_transform_z()
}
\description{
Rcmdr window for z transformation
}
\seealso{
Other transformations: \code{\link{window_num_transform_log}}
}
\concept{transformations}
\keyword{internal}
