% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdr__set_contrasts.R
\name{window_set_contrasts}
\alias{window_set_contrasts}
\title{Set contrasts for a factor variable (Rcmdr-based)}
\usage{
window_set_contrasts()
}
\description{
Based on Rcmdr function `setContrasts()`.
}
\keyword{internal}
