% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_z_transform_2.R
\name{window_z_transform_2}
\alias{window_z_transform_2}
\title{Rcmdr window for z transformation (does not work)}
\usage{
window_z_transform_2()
}
\description{
Rcmdr window for z transformation (does not work)
}
\seealso{
Other transformations: \code{\link{window_log_transform}},
  \code{\link{window_z_transform}}
}
\keyword{internal}
