% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_entry.R
\name{bs_entry}
\alias{bs_entry}
\alias{get_values.bs_entry}
\alias{set_values.bs_entry}
\alias{tk_get_state.bs_entry}
\alias{tk_disable.bs_entry}
\alias{tk_normalize.bs_entry}
\alias{tk_activate.bs_entry}
\title{TCL/TK text entry widget.}
\usage{
bs_entry(parent = top, width = "28", value = "", label = "",
  label_position = c("left", "above", "right", "none"),
  label_color = getRcmdr("title.color"), padx = 0, pady = 0,
  sticky = "w", sticky_label = sticky, sticky_text = sticky,
  main_frame = tk2frame(parent), text_frame = tk2frame(main_frame),
  label_frame = tk2frame(main_frame), tip = "", label_tip = "",
  scroll_x = FALSE, on_click = do_nothing,
  on_double_click = do_nothing, on_triple_click = do_nothing,
  on_release = do_nothing, on_click_3 = do_nothing,
  on_double_click_3 = do_nothing, on_triple_click_3 = do_nothing,
  on_release_3 = do_nothing, on_key_release = do_nothing,
  use_context_menu = TRUE, bind_clear = TRUE,
  variable = tclVar(value), ...)

\method{get_values}{bs_entry}(obj, ..., trim = TRUE)

\method{set_values}{bs_entry}(obj, values, ..., add = FALSE)

\method{tk_get_state}{bs_entry}(obj, ...)

\method{tk_disable}{bs_entry}(obj, ..., foreground = "grey")

\method{tk_normalize}{bs_entry}(obj, ...,
  foreground = getRcmdr("title.color"))

\method{tk_activate}{bs_entry}(obj, ...,
  foreground = getRcmdr("title.color"))
}
\arguments{
\item{trim}{(logical) Should leading and tailing whitespace be trimmed in the result?}

\item{add}{(logical) Should values be added to currently existing ones?}

\item{foreground}{(string) Label text color.}
}
\description{
TCL/TK text entry widget.
}
\keyword{internal}
