% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_listbox.R
\name{bs_listbox}
\alias{bs_listbox}
\alias{get_size.listbox}
\alias{get_size.tkwin}
\alias{get_values_listbox}
\alias{get_values.listbox}
\alias{get_values.tk2listbox}
\alias{set_values_listbox}
\alias{set_values.listbox}
\alias{set_values.tk2listbox}
\alias{set_values.listbox_with_filter}
\alias{get_selection_ind_listbox}
\alias{get_selection_listbox}
\alias{get_selection.listbox}
\alias{get_selection.tk2listbox}
\alias{get_selection_length.listbox}
\alias{get_selection_length.tk2listbox}
\alias{set_selection_listbox}
\alias{set_selection.listbox}
\alias{set_selection.tk2listbox}
\alias{add_selection.listbox}
\alias{add_selection.tk2listbox}
\alias{move_selected_row_in_listbox}
\alias{bind_row_swap.listbox}
\alias{bind_row_swap.tk2listbox}
\alias{set_yview.listbox}
\alias{set_yview.tk2listbox}
\alias{tk_see.listbox}
\alias{tk_see.tk2listbox}
\alias{tk_get_state.listbox}
\alias{tk_get_state.tk2listbox}
\alias{tk_disable.listbox}
\alias{tk_disable.tk2listbox}
\alias{tk_normalize.listbox}
\alias{tk_normalize.tk2listbox}
\title{List box widget (TCL/TK)}
\usage{
bs_listbox(parent, values = variables_all(), value = NULL,
  selection = NULL, selectmode = c("single", "extended", "browse",
  "multiple"), title = NULL, subtitle = NULL, tip = "",
  height = getRcmdr("variable.list.height"),
  width = getRcmdr("variable.list.width"), enabled = TRUE,
  scroll = c("both", "x", "y", "none"), autoscroll = c("x", "y",
  "both", "none"), use_filter = FALSE, filter_label = "Filter",
  sticky = "nw", on_select = do_nothing, on_click = do_nothing,
  on_double_click = do_nothing, on_triple_click = do_nothing,
  on_release = do_nothing, on_click_3 = do_nothing,
  on_double_click_3 = do_nothing, on_triple_click_3 = do_nothing,
  on_release_3 = do_nothing, on_keyboard = c("select", "scroll",
  "ignore"), on_keyboard_fun = do_nothing, bind_row_swap = FALSE,
  title_sticky = "w", subtitle_sticky = title_sticky, ...)

\method{get_size}{listbox}(obj, ...)

\method{get_size}{tkwin}(obj, ...)

get_values_listbox(listbox)

\method{get_values}{listbox}(obj, ...)

\method{get_values}{tk2listbox}(obj, vals, ...)

set_values_listbox(listbox, values, clear = TRUE)

\method{set_values}{listbox}(obj, values, ..., clear = TRUE)

\method{set_values}{tk2listbox}(obj, values, ..., clear = TRUE)

\method{set_values}{listbox_with_filter}(obj, values, ..., clear = TRUE)

get_selection_ind_listbox(listbox)

get_selection_listbox(listbox)

\method{get_selection}{listbox}(obj, ...)

\method{get_selection}{tk2listbox}(obj, ...)

\method{get_selection_length}{listbox}(obj, ...)

\method{get_selection_length}{tk2listbox}(obj, ...)

set_selection_listbox(listbox, sel, clear = TRUE)

\method{set_selection}{listbox}(obj, sel, clear = TRUE, ...)

\method{set_selection}{tk2listbox}(obj, sel, clear = TRUE, ...)

\method{add_selection}{listbox}(obj, sel, ...)

\method{add_selection}{tk2listbox}(obj, sel, ...)

move_selected_row_in_listbox(listbox, move_to = "")

bind_row_swap.listbox(obj, ...)

bind_row_swap.tk2listbox(obj, ...)

\method{set_yview}{listbox}(obj, ind, ...)

\method{set_yview}{tk2listbox}(obj, ind, ...)

\method{tk_see}{listbox}(obj, ind, ...)

\method{tk_see}{tk2listbox}(obj, ind, ...)

\method{tk_get_state}{listbox}(obj, ...)

\method{tk_get_state}{tk2listbox}(obj, ...)

\method{tk_disable}{listbox}(obj, ..., background = "grey95")

\method{tk_disable}{tk2listbox}(obj, ..., background = "grey95")

\method{tk_normalize}{listbox}(obj, ..., background = "white")

\method{tk_normalize}{tk2listbox}(obj, ..., background = "white")
}
\arguments{
\item{parent}{Parent TCL/TK frame.}

\item{values}{(character) Vector of possible values.}

\item{value}{(character or NULL) Values that should be initially selected.}

\item{selection}{(integer) Numeric indices of initially selected lines.
Can be used instead of \code{value}.}

\item{selectmode}{("single", "extended", "browse", "multiple") Selection mode.}

\item{title}{(string) Title.}

\item{subtitle}{(string) Subtitle.}

\item{tip}{(string) Message that is displayed on muse hover.}

\item{height}{(integer) Height of list box in lines.}

\item{width}{(one or two integers) Minimum and maximum width of list box.}

\item{enabled}{(logical) If \code{TRUE}, widget is enabled by default.}

\item{scroll}{("both", "x", "y", "none") Do we add scrollbars?
Indicates scrollbar position.}

\item{autoscroll}{("x", "y", "both", "none")
Do we automatically hide scrollbars if not needed?
Indicates scrollbar position.}

\item{use_filter}{(logical) Should list box values filter box be displayed.}

\item{filter_label}{(string) Label for filter entry box.}

\item{sticky}{(combination of "n", "e", "w", "s" or "").}

\item{on_select}{Function that is activated on value selection.}

\item{on_click}{Function that is activated on single left-click.}

\item{on_double_click}{on double left-click.}

\item{on_triple_click}{Function that is activated on triple left-click.}

\item{on_release}{Function that is activated on release of left-click.}

\item{on_click_3}{Function that is activated on single right-click.}

\item{on_double_click_3}{Function that is activated on double right-click.}

\item{on_triple_click_3}{Function that is activated on triple right-click.}

\item{on_release_3}{Function that is activated on release of right-click.}

\item{on_keyboard}{("select", "scroll", "ignore") Action on keyboard letter click:
\describe{
    \item{"select"}{toggle selection of values that start with the pressed letter.
    This letter binding is good for single-selection boxes, otherwise selection should not be toggled (see "scroll")}
    \item{"scroll"}{toggle view of values that start with the pressed letter.
     Do not toggle selection (i.e., do not select or deselect). }
    \item{"ignore"}{Do not bind any action. Letter binding is good for read-only list boxes only.}
}}

\item{on_keyboard_fun}{Function that is activated on keyboard button press, if
\code{on_keyboard} is "select" or "scroll".}

\item{bind_row_swap}{(Does not work yet!) if TRUE, Ctrl/Alt + Up/Down keys move rows in the list box}

\item{title_sticky}{(combination of "n", "e", "w", "s" or "")}

\item{subtitle_sticky}{(combination of "n", "e", "w", "s" or "")}

\item{...}{Additional arguments.  \enumerate{
   \item For \code{bs_listbox()}: arguments passed to function \code{\link[tcltk2]{tk2listbox}()}.
   \item For other functions: arguments passed to further methods.
}}

\item{obj}{Widget.}

\item{listbox}{List box widget.}

\item{vals}{Values.}

\item{clear}{(logical) Should values be cleared?}

\item{sel}{Selection. Usually either character values or numeric indices.}

\item{move_to}{(string) Action:
\itemize{
   \item "top" - selected row is moved to the first row.
   \item "-1"  - position is decreased by 1 row.
   \item "+1"  - position inreased by 1row.
   \item "end" - selected row is moved to the last row.
}}

\item{ind}{Index. Usually either character values or numeric indices.}

\item{background}{(string) background color.}

\item{values}{Values.}
}
\description{
A list box with constant number of rows.
}
\examples{

\dontrun{\donttest{

# Active dataset must be selected
top <- tktoplevel()
lb1 <- bs_listbox(top, values = LETTERS, value = "K")
lb2 <- bs_listbox(top, values = LETTERS, selection = c(12, 15), selectmode = "multiple")
tkgrid(lb1$frame, lb2$frame)

}}

}
\seealso{
\code{\link[tcltk2]{tk2listbox}()}, \cr
         \url{https://www.tcl.tk/man/tcl8.4/TkCmd/listbox.htm}
}
