% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_radiobuttons.R
\name{bs_radiobuttons}
\alias{bs_radiobuttons}
\title{Radiobuttons widget.}
\usage{
bs_radiobuttons(parent = top, buttons, value = buttons[1],
  title = NULL, variable = NULL, labels = NULL, commands = list(),
  default_command = do_nothing, tips = list(), default_tip = "",
  border = FALSE, layout = c("vertical", "horizontal"),
  sticky_buttons = "w", sticky_title = "w")
}
\arguments{
\item{parent}{Parent frame.}

\item{buttons}{(vector of strings) Value names for each radiobutton.}

\item{value}{(string) Initial value. One of `buttons` values.}

\item{title}{Title for the set of radiobuttons.}

\item{variable}{Tcl/Tk variable. Is `NULL`, a new variable will be created.}

\item{labels}{(vector of strings) Labels for each radiobutton.}

\item{commands}{A named list of commands (functions) for radiobutton. The names must match the values of "buttons".}

\item{default_command}{(function) A default command.}

\item{tips}{A named list of strings to be used as tips for radiobutton.
The names must match the values of "buttons".}

\item{default_tip}{(string) a default tip.}

\item{border}{(logical) Flag if the frame should have a border.}

\item{layout}{(string) One of "vertical" (default) and "horizontal".}

\item{sticky_buttons}{(string) `sticky` option for buttons.}

\item{sticky_title}{(string) `sticky` option for title (if no border is used).}
}
\value{
A named list with fields `frame` (frame with the radiobuttons),
 `var` (tcl/tk variables for each box),
 and `obj` (tcl/tk objects for each box).
}
\description{
Radiobuttons widget with command functions and tips when mouse is over the box.
}
\examples{
\dontrun{\donttest{

top <- tcltk::tktoplevel()

buttons_1 <- bs_radiobuttons(top, c("A", "B", "C"))
tcltk::tkgrid(buttons_1$frame)


top <- tcltk::tktoplevel()
buttons_2 <- bs_radiobuttons(top, buttons = c("A", "B", "C"),
                             border = TRUE, title = "Buttons")
tcltk::tkgrid(buttons_2$frame)


top <- tcltk::tktoplevel()
buttons_3 <- bs_radiobuttons(top, c("A", "B", "C"), layout = "h",
                         title = "Buttons", sticky_buttons = "")
tcltk::tkgrid(buttons_3$frame)

}}
}
