% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{characterP}
\alias{characterP}
\title{characterP
Return TRUE, if at least n character variables exist in the active dataset.}
\usage{
characterP(n = 1)
}
\arguments{
\item{n}{Minimum number of character variables}
}
\description{
characterP
Return TRUE, if at least n character variables exist in the active dataset.
}
\keyword{internal}
