% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{export_to_clipboard}
\alias{export_to_clipboard}
\alias{export_to_clipboard_active_ds_tab}
\title{Expotr text/dataframe to clipboard.}
\usage{
export_to_clipboard(ds_name = active_dataset_0(), sep = ",")

export_to_clipboard_active_ds_tab()
}
\arguments{
\item{ds_name}{(string) Dataset's name.}

\item{sep}{(string) Data field separator, if data frame is exported.}
}
\description{
Expotr text/dataframe to clipboard.
}
