% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{extract-fileparts}
\alias{extract-fileparts}
\alias{extract_filename}
\alias{do_nothing}
\alias{is_url}
\alias{is_url_accessible}
\title{Extract file parts.}
\usage{
extract_filename(str)

do_nothing(...)

is_url(str)

is_url_accessible(str)
}
\arguments{
\item{str}{(character) Path to file (with filename and extension).}
}
\description{
Extract file parts.
}
\keyword{internal}
