% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{extract_path}
\alias{extract_path}
\alias{extract-fileparts}
\alias{extract_filename}
\alias{extract_extension}
\alias{path_ext_set_2}
\alias{is_url}
\title{Extract file parts.}
\usage{
extract_path(str)

extract_filename(str)

extract_extension(str)

path_ext_set_2(path, ext)

is_url(str)
}
\arguments{
\item{str}{(character) Path to file (with filename and extension).}
}
\description{
Extract file parts.
}
\keyword{internal}
