% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command_buttons_mode.R,
%   R/generic_functions.R, R/helper-functions.R, R/tcltk__bs_check_boxes.R,
%   R/tcltk__variableListBox2.R, R/tcltk_commands.R, R/tcltk_elements.R,
%   R/tcltk_listbox.R, R/tcltk_ok_cancel_help__macro.R,
%   R/tk_widget_modify_names.R, R/tk_widget_y_gr_boxes.R
\name{command_buttons_mode_rcmdr}
\alias{command_buttons_mode_rcmdr}
\alias{command_buttons_mode_biostat}
\alias{selectActiveDataSet}
\alias{selectActiveModel}
\alias{tk_get_state}
\alias{tk_activate}
\alias{tk_disable}
\alias{tk_normalize}
\alias{get_size}
\alias{get_values}
\alias{set_values}
\alias{get_selection}
\alias{get_selection_length}
\alias{set_selection}
\alias{set_new_selection}
\alias{set_yview}
\alias{Helper-functions}
\alias{get_active_ds}
\alias{list_objects_of_class}
\alias{var_pos_n}
\alias{list_summaries_Models}
\alias{eval_glue}
\alias{eval_}
\alias{eval_text}
\alias{style_cmd}
\alias{try_command}
\alias{gettext_ezr}
\alias{gettext_bs}
\alias{s2u}
\alias{u2s}
\alias{spaces}
\alias{logger_error}
\alias{tk_label_blue}
\alias{swap}
\alias{swap_rows}
\alias{correct_row_index}
\alias{clean_str}
\alias{function_not_implemented}
\alias{unique_obj_names}
\alias{unique_df_name}
\alias{unique_colnames}
\alias{unique_colnames_2}
\alias{variables_with_unique_values}
\alias{variables_with_unique_values_P}
\alias{path_truncate}
\alias{show_error_messages}
\alias{is_not_valid_name}
\alias{is_empty_name}
\alias{is_not_empty_name}
\alias{variable_is_not_selected}
\alias{are_not_valid_names}
\alias{forbid_to_replace_variable}
\alias{forbid_to_replace_variables}
\alias{forbid_to_replace_object}
\alias{show_code_evaluation_error_message}
\alias{nonFactorsP}
\alias{class_ggplot_P}
\alias{list_objects_ggplot}
\alias{objects_in_env_P}
\alias{command_use_1_window_rcmdr}
\alias{command_use_3_window_rcmdr}
\alias{command_use_rcmdr_windows}
\alias{bs_check_boxes}
\alias{variableListBox2}
\alias{cursor_set_busy}
\alias{cursor_set_idle}
\alias{tclvalue_lgl}
\alias{tclvalue_int}
\alias{tclvalue_int_split}
\alias{tclvalue_chr}
\alias{tclvalue_chr_split}
\alias{label_rcmdr}
\alias{labeled_frame}
\alias{tk_title}
\alias{radioButtons_horizontal}
\alias{radiobuttons_env}
\alias{bs_listbox}
\alias{listbox_get_values}
\alias{listbox_set_values}
\alias{listbox_get_selection_ind}
\alias{listbox_get_selection}
\alias{listbox_set_selection}
\alias{listbox_set_new_selection}
\alias{get_selection.listbox}
\alias{get_selection_length.listbox}
\alias{get_size.listbox}
\alias{get_size.tkwin}
\alias{set_selection.listbox}
\alias{set_new_selection.listbox}
\alias{get_values.listbox}
\alias{set_values.listbox}
\alias{set_yview.listbox}
\alias{tk_disable.listbox}
\alias{tk_normalize.listbox}
\alias{tk_get_state.listbox}
\alias{ok_cancel_help}
\alias{tk_widget_modify_names}
\alias{print.bs_tk_widget}
\alias{tk_widget_boxes_y_gr}
\title{Helper functions for RcmdrPlugin.biostat.}
\usage{
command_buttons_mode_rcmdr()

command_buttons_mode_biostat()

selectActiveDataSet()

selectActiveModel()

tk_get_state(obj, ...)

tk_activate(obj, ...)

tk_disable(obj, ...)

tk_normalize(obj, ...)

get_size(obj, ...)

get_values(obj, ...)

set_values(obj, ...)

get_selection(obj, ...)

get_selection_length(obj, ...)

set_selection(obj, sel, ...)

set_new_selection(obj, sel, ...)

set_yview(obj, ...)

get_active_ds()

list_objects_of_class(class = NULL, all.names = FALSE,
  envir = parent.frame())

var_pos_n(variables, type = c("all", "numeric", "num", "factor",
  "fct_like", "factor_strict", "fct", "twoLevelFactor", "fct2",
  "nonfactor", "nfct", "character", "chr", "logical", "lgl"),
  vars = NULL)

list_summaries_Models(envir = .GlobalEnv, ...)

eval_glue(..., envir = parent.frame(), .sep = "", .open = "{",
  .close = "}", envir_eval = envir, envir_glue = envir)

eval_(x, envir = parent.frame(), ...)

eval_text(x, envir = parent.frame(), ...)

style_cmd(command, indent_by = 4, ...)

try_command(x)

gettext_ezr(...)

gettext_bs(...)

s2u(str)

u2s(str)

spaces(n, symbol = " ")

logger_error(command = NULL, error_msg = NULL)

tk_label_blue(...)

swap(x, i, j)

swap_rows(x, i, j)

correct_row_index(i, n_max)

clean_str(str, ...)

function_not_implemented()

unique_obj_names(names = ActiveDataSet(), preffix = "", suffix = "",
  list_of_choices = objects(all.names = TRUE, envir = .GlobalEnv),
  all_numbered = FALSE)

unique_df_name(names = ActiveDataSet(), preffix = "", suffix = "",
  list_of_choices = objects(all.names = TRUE, envir = .GlobalEnv),
  all_numbered = FALSE)

unique_colnames(names = "", preffix = "", suffix = "",
  list_of_choices = listVariables(), all_numbered = FALSE)

unique_colnames_2(names = "", preffix = "", suffix = "",
  list_of_choices = listVariables(), all_numbered = TRUE)

variables_with_unique_values()

variables_with_unique_values_P(n = 1)

path_truncate(path, max_length = 30)

show_error_messages(message, popup_msg = message, title = "Error")

is_not_valid_name(name)

is_empty_name(name)

is_not_empty_name(name)

variable_is_not_selected(obj, obj_type = "variable")

are_not_valid_names(name)

forbid_to_replace_variable(name)

forbid_to_replace_variables(name)

forbid_to_replace_object(name, envir = .GlobalEnv)

show_code_evaluation_error_message()

nonFactorsP(n = 1)

class_ggplot_P(n = 1)

list_objects_ggplot(envir = .GlobalEnv)

objects_in_env_P(n = 1, envir = .GlobalEnv, ...)

command_use_1_window_rcmdr()

command_use_3_window_rcmdr()

command_use_rcmdr_windows(console.output = NULL)

bs_check_boxes(window = top, frame = stop("frame not supplied"),
  boxes = stop("boxes not supplied"), initialValues = NULL,
  labels = stop("labels not supplied"), title = NULL, ttk = FALSE,
  commands = list())

variableListBox2(parentWindow, variableList = Variables(),
  bg = "white", selectmode = "single", export = "FALSE",
  initialSelection = NULL,
  listHeight = getRcmdr("variable.list.height"),
  listWidth = getRcmdr("variable.list.width"), onClick_fun = function()
  { }, onDoubleClick_fun = function() { },
  onTripleClick_fun = function() { }, onRelease_fun = function() { },
  onClick3_fun = function() { }, onDoubleClick3_fun = function() { },
  onTripleClick3_fun = function() { }, onRelease3_fun = function() { },
  title)

cursor_set_busy(frame = NULL)

cursor_set_idle(frame = NULL)

tclvalue_lgl(x)

tclvalue_int(x)

tclvalue_int_split(x)

tclvalue_chr(x, trim = TRUE, ...)

tclvalue_chr_split(x)

label_rcmdr(..., fg = NULL)

labeled_frame(parent, label = NULL, ...)

tk_title(parent = top, text = "xxx_title", pady = c(5, 9),
  font = tkfont.create(weight = "bold", size = 9), ...)

radioButtons_horizontal(window = top, name = stop("name not supplied"),
  buttons = stop("buttons not supplied"), values = NULL,
  initialValue = ..values[1], labels = stop("labels not supplied"),
  title = NULL, title.color = NULL, right.buttons = FALSE,
  command = function() { }, sticky_title = "w", sticky_buttons = "e")

radiobuttons_env(window = top, name = stop("name not supplied"),
  buttons = stop("buttons not supplied"), values = NULL,
  initialValue = ..values[1], labels = stop("labels not supplied"),
  title = "", title.color = getRcmdr("title.color"),
  right.buttons = FALSE, command = function() { },
  env = parent.frame())

bs_listbox(parentWindow, variableList = variables_all(), bg = "white",
  selectmode = "single", export = "FALSE", initial_selection = NULL,
  listHeight = getRcmdr("variable.list.height"),
  listWidth = getRcmdr("variable.list.width"), on_click = function() {
  }, on_double_click = function() { }, on_triple_click = function() {
  }, on_release = function() { }, on_click_3 = function() { },
  on_double_click_3 = function() { }, on_triple_click_3 = function() {
  }, on_release_3 = function() { }, title)

listbox_get_values(listbox)

listbox_set_values(listbox, variableList)

listbox_get_selection_ind(listbox)

listbox_get_selection(listbox)

listbox_set_selection(listbox, sel, clear = FALSE)

listbox_set_new_selection(listbox, sel)

\method{get_selection}{listbox}(obj, ...)

\method{get_selection_length}{listbox}(obj, ...)

\method{get_size}{listbox}(obj, ...)

\method{get_size}{tkwin}(obj, ...)

\method{set_selection}{listbox}(obj, sel, clear = FALSE, ...)

\method{set_new_selection}{listbox}(obj, sel, ...)

\method{get_values}{listbox}(obj, vals, ...)

\method{set_values}{listbox}(obj, vals, ...)

\method{set_yview}{listbox}(obj, ind, ...)

\method{tk_disable}{listbox}(obj, ..., background = "grey95")

\method{tk_normalize}{listbox}(obj, ..., background = "white")

\method{tk_get_state}{listbox}(obj)

ok_cancel_help(window = top, helpSubject = NULL, model = FALSE,
  reset = NULL, apply = NULL, helpPackage = NULL, sticky = "w")

tk_widget_modify_names(parent = top,
  init_val_radiobuttons = c("overwrite", "modify"),
  init_val_checkbox = "0", init_val_prefix = "",
  init_val_suffix = "", cmd_radiobuttons = function() { },
  cmd_checkbox = function() { })

\method{print}{bs_tk_widget}(x, ...)

tk_widget_boxes_y_gr(parent_frame = top, y_title = title_var_1,
  y_var_type = NA_character_, y_vars = NULL, y_initial = NULL,
  y_select_mode = "single", y_params = list(),
  gr_title = title_gr_0_n, gr_var_type = NA_character_,
  gr_vars = NULL, gr_initial = NULL, gr_select_mode = "multiple",
  gr_params = list(), ch_label = "Use groups", ch_initial = "0",
  ch_commands = function() { }, ch_params = list(), listHeight = 7,
  listWidth = c(25, Inf), add_to_grid = TRUE)
}
\arguments{
\item{...}{other arguments}

\item{str}{sring}

\item{frame}{Tcl/Tk frame object.}
}
\description{
Helper functions for package \pkg{RcmdrPlugin.biostat}.
}
\keyword{internal}
