% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defmacro.R, R/helper-functions.R,
%   R/ok_cancel_help__macro.R, R/tcltk_elements.R
\name{defmacro}
\alias{defmacro}
\alias{Helper-functions}
\alias{get_active_ds}
\alias{list_objects_of_class}
\alias{list_summaries_Models}
\alias{eval_glue}
\alias{eval_}
\alias{style_cmd}
\alias{gettext_EZR}
\alias{gettext_Bio}
\alias{s2u}
\alias{u2s}
\alias{spaces}
\alias{logger_error}
\alias{tk_label_blue}
\alias{swap}
\alias{swap_rows}
\alias{correct_row_index}
\alias{function_not_implemented}
\alias{unique_obj_names}
\alias{unique_df_name}
\alias{unique_colnames}
\alias{unique_colnames_2}
\alias{variables_with_unique_values}
\alias{variables_with_unique_values_P}
\alias{path_truncate}
\alias{show_error_messages}
\alias{is_not_valid_name}
\alias{is_empty_name}
\alias{is_not_empty_name}
\alias{forbid_to_replace_variable}
\alias{forbid_to_replace_object}
\alias{show_code_evaluation_error_message}
\alias{nonFactorsP}
\alias{class_ggplot_P}
\alias{list_objects_ggplot}
\alias{objects_in_env_P}
\alias{ok_cancel_help}
\alias{label_rcmdr}
\alias{labeled_frame}
\alias{variableListBox2}
\alias{bs_check_boxes}
\alias{radioButtons_horizontal}
\alias{radiobuttons_env}
\title{Helper functions for RcmdrPlugin.biostat.}
\usage{
defmacro(..., expr)

get_active_ds()

list_objects_of_class(class = NULL, all.names = FALSE,
  envir = parent.frame())

list_summaries_Models(envir = .GlobalEnv, ...)

eval_glue(..., envir = parent.frame(), .sep = "", .open = "{",
  .close = "}", envir_eval = envir, envir_glue = envir)

eval_(x, envir = parent.frame(), ...)

style_cmd(command, indent_by = 4, ...)

gettext_EZR(...)

gettext_Bio(...)

s2u(str)

u2s(str)

spaces(n, symbol = " ")

logger_error(command)

tk_label_blue(...)

swap(x, i, j)

swap_rows(x, i, j)

correct_row_index(i, n_max)

function_not_implemented()

unique_obj_names(names = ActiveDataSet(), preffix = "", suffix = "",
  list_of_choices = objects(all.names = TRUE, envir = .GlobalEnv),
  all_numbered = FALSE)

unique_df_name(names = ActiveDataSet(), preffix = "", suffix = "",
  list_of_choices = objects(all.names = TRUE, envir = .GlobalEnv),
  all_numbered = FALSE)

unique_colnames(names = "", preffix = "", suffix = "",
  list_of_choices = listVariables(), all_numbered = FALSE)

unique_colnames_2(names = "", preffix = "", suffix = "",
  list_of_choices = listVariables(), all_numbered = TRUE)

variables_with_unique_values()

variables_with_unique_values_P(n = 1)

path_truncate(path, max_length = 30)

show_error_messages(message, message2 = message, title = "")

is_not_valid_name(name)

is_empty_name(name)

is_not_empty_name(name)

forbid_to_replace_variable(name)

forbid_to_replace_object(name, envir = .GlobalEnv)

show_code_evaluation_error_message()

nonFactorsP(n = 1)

class_ggplot_P(n = 1)

list_objects_ggplot(envir = .GlobalEnv)

objects_in_env_P(n = 1, envir = .GlobalEnv, ...)

ok_cancel_help(window = top, helpSubject = NULL, model = FALSE,
  reset = NULL, apply = NULL, helpPackage = NULL)

label_rcmdr(..., fg = NULL)

labeled_frame(parent, label = NULL, ...)

variableListBox2(parentWindow, variableList = Variables(),
  bg = "white", selectmode = "single", export = "FALSE",
  initialSelection = NULL,
  listHeight = getRcmdr("variable.list.height"),
  onClick_fun = function() { }, onDoubleClick_fun = function() { },
  onTripleClick_fun = function() { }, onRelease_fun = function() { },
  onClick3_fun = function() { }, onDoubleClick3_fun = function() { },
  onTripleClick3_fun = function() { }, onRelease3_fun = function() { },
  title)

bs_check_boxes(window = top, frame = stop("frame not supplied"),
  boxes = stop("boxes not supplied"), initialValues = NULL,
  labels = stop("labels not supplied"), title = NULL, ttk = FALSE,
  commands = list())

radioButtons_horizontal(window = top, name = stop("name not supplied"),
  buttons = stop("buttons not supplied"), values = NULL,
  initialValue = ..values[1], labels = stop("labels not supplied"),
  title = NULL, title.color = NULL, right.buttons = FALSE,
  command = function() { }, sticky_title = "w", sticky_buttons = "e")

radiobuttons_env(window = top, name = stop("name not supplied"),
  buttons = stop("buttons not supplied"), values = NULL,
  initialValue = ..values[1], labels = stop("labels not supplied"),
  title = "", title.color = getRcmdr("title.color"),
  right.buttons = FALSE, command = function() { },
  env = parent.frame())
}
\description{
Helper functions for package \pkg{RcmdrPlugin.biostat}.
}
\keyword{internal}
