% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{is_valid_name}
\alias{is_valid_name}
\title{Check if name is valid}
\usage{
is_valid_name(name, parent = CommanderWindow())
}
\arguments{
\item{name}{(character) a single string.}

\item{parent}{(tkwin object) a parent Tcl/Tk window.}
}
\value{
Logical value TRUE if string meets requirements, FALSE - otherwise.
}
\description{
Function checks if string meets requirements for a single, valid, non-empty
R variable name. In case of incalid name, a box wit error message is thrown,
error message is written to R Commander window and result FALSE is returned.
Oterwise, TRUE is returned.\cr
\code{is_valit_name} - checks character validity.\cr
\code{is_empty_name} - does not check character validity.
}
\examples{
\dontrun{\donttest{
is_valid_name("a")
is_valid_name("")
is_valid_name("|||")

is_empty_name("a")
is_empty_name("")
is_empty_name("|||")
}}
}
\keyword{internal}
