% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{is_valid_name}
\alias{is_valid_name}
\title{Check if name is valid}
\usage{
is_valid_name(name)
}
\arguments{
\item{name}{(character) a single string.}
}
\value{
Logical value TRUE if string meets requirements, FALSE - otherwise.
}
\description{
Function checks if string meets requirements for a single, valid, non-empty
R variable name. In case of incalid name, a box wit error message is thrown,
error message is written to R Commander window and result FALSE is returned.
Oterwise, TRUE is returned.
}
\examples{
is_valid_name("a")
is_valid_name("")
is_valid_name("|||")
}
\keyword{internal}
