% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_biostat_mode.R
\name{load_rcmdr_biostat_mode}
\alias{load_rcmdr_biostat_mode}
\title{Load Biostat mode in R Commander}
\usage{
load_rcmdr_biostat_mode()
}
\description{
Restarts R Commander (to load plugins) and sets Biostat mode in R Commander.
Packages \pkg{Rcmdr} and \pkg{RcmdrPlugin.biostat} must be attached.
For this purpose `library("RcmdrPlugin.biostat")` is sufficient.
If restart is not needed, `set_biostat_mode()` is sufficient.
}
\examples{
\dontrun{\donttest{

library("RcmdrPlugin.biostat")
load_rcmdr_biostat_mode()

}}
}
