% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R, R/menu-active-dataset.R,
%   R/menu-conf_interval.R, R/menu-data.R, R/menu-export.R, R/menu-import.R,
%   R/menu-models.R, R/menu-plots.R, R/menu-rows.R, R/menu-settings.R,
%   R/menu-summary.R, R/menu-variables.R, R/window_anova_kw_mood_tests.R,
%   R/window_col_compute.R, R/window_dataset_bind.R,
%   R/window_dataset_bind_cols.R, R/window_dataset_bind_rows.R,
%   R/window_dataset_join.R, R/window_do_summary.R, R/window_pander.R,
%   R/window_qqplot_biostat.R, R/window_test_normality.R, R/window_tmp.R
\name{function_not_implemented}
\alias{function_not_implemented}
\alias{class_gglot_P}
\alias{objects_in_env_P}
\alias{Menu-window-functions}
\alias{command_dataset_refresh}
\alias{command_dataset_class}
\alias{command_colnames}
\alias{command_dataset_print}
\alias{command_dataset_pander}
\alias{command_dataset_view}
\alias{command_dataset_dim}
\alias{command_dataset_as_df}
\alias{command_dataset_as_tibble}
\alias{command_dataset_as_dt}
\alias{window_ci_mean}
\alias{window_ci_mean_diff}
\alias{window_ci_median}
\alias{window_ci_prop_multi}
\alias{window_ci_prop_binom}
\alias{window_ci_boot}
\alias{Menu-window-functions}
\alias{window_export_as_textfile}
\alias{window_export_as_rds}
\alias{window_import_rds}
\alias{dataset_import_excel}
\alias{command_std_lm_coeffs}
\alias{command_get_model_class}
\alias{command_model_print}
\alias{open_new_plots_window}
\alias{window_easyPlot}
\alias{window_ggplotgui}
\alias{window_show_colors}
\alias{window_colourPicker}
\alias{window_histogram}
\alias{window_density_plot}
\alias{window_plot_dot_err}
\alias{window_plot_bar_err}
\alias{command_rownames}
\alias{command_rowid_to_col}
\alias{command_rownames_to_col}
\alias{command_col_to_rownames}
\alias{window_rows_arrange}
\alias{window_rows_filter}
\alias{window_rows_slice}
\alias{command_load_packages}
\alias{command_getwd}
\alias{command_get_locale}
\alias{command_set_locale_lt}
\alias{command_set_locale_en}
\alias{command_set_locale_ru}
\alias{command_set_locale_default}
\alias{command_session_info_devtools}
\alias{command_session_info_utils}
\alias{command_glimpse}
\alias{command_summary}
\alias{summary_head_tail}
\alias{summary_head}
\alias{summary_tail}
\alias{summary_Hmisc_describe}
\alias{summary_ds_screener}
\alias{summary_psych_describe}
\alias{summary_psych_describeBy}
\alias{command_all_chr_to_fctr}
\alias{window_anova_kw_mood_tests}
\alias{window_col_compute}
\alias{window_dataset_bind}
\alias{window_dataset_bind_cols}
\alias{window_dataset_bind_rows}
\alias{window_dataset_join}
\alias{window_do_summary}
\alias{window_pander}
\alias{window_qqplot_biostat}
\alias{window_normality_test}
\alias{window_tmp}
\title{RcmdrPlugin.biostat functions for menus and windows.}
\usage{
function_not_implemented(x = NULL)

class_gglot_P(n = 1)

objects_in_env_P(n = 1, ...)

command_dataset_refresh()

command_dataset_class()

command_colnames()

command_dataset_print()

command_dataset_pander()

command_dataset_view()

command_dataset_dim()

command_dataset_as_df()

command_dataset_as_tibble()

command_dataset_as_dt()

window_ci_mean()

window_ci_mean_diff()

window_ci_median()

window_ci_prop_multi()

window_ci_prop_binom()

window_ci_boot()

window_export_as_textfile()

window_export_as_rds()

window_import_rds()

dataset_import_excel()

command_std_lm_coeffs()

command_get_model_class()

command_model_print()

open_new_plots_window()

window_easyPlot()

window_ggplotgui()

window_show_colors()

window_colourPicker()

window_histogram()

window_density_plot()

window_plot_dot_err()

window_plot_bar_err()

command_rownames()

command_rowid_to_col()

command_rownames_to_col()

command_col_to_rownames()

window_rows_arrange()

window_rows_filter()

window_rows_slice()

command_load_packages()

command_getwd()

command_get_locale()

command_set_locale_lt()

command_set_locale_en()

command_set_locale_ru()

command_set_locale_default()

command_session_info_devtools()

command_session_info_utils()

command_glimpse()

command_summary()

summary_head_tail()

summary_head()

summary_tail()

summary_Hmisc_describe()

summary_ds_screener()

summary_psych_describe()

summary_psych_describeBy()

command_all_chr_to_fctr()

window_anova_kw_mood_tests()

window_col_compute()

window_dataset_bind()

window_dataset_bind_cols()

window_dataset_bind_rows()

window_dataset_join()

window_do_summary()

window_pander()

window_qqplot_biostat()

window_normality_test()

window_tmp()
}
\description{
Functions that open Rcmdr menus and windows windows.

Functions that open Rcmdr menus and windows windows.
}
\keyword{internal}
