% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_dataset.R, R/menu-ci.R, R/menu-data.R,
%   R/menu-export.R, R/menu-import.R, R/menu-models.R, R/menu-plots.R,
%   R/menu-rows.R, R/menu-settings.R, R/menu-summary.R, R/menu-variables.R,
%   R/online_tools.R, R/window_ci_mean.R, R/window_dataset_bind_cols.R,
%   R/window_dataset_bind_rows.R, R/window_dataset_class.R,
%   R/window_dataset_join.R, R/window_dataset_print_as_md.R,
%   R/window_dataset_select.R, R/window_export_to_excel.R,
%   R/window_fct_drop_lvls.R, R/window_fct_lvls_reorder_manual.R,
%   R/window_image_digitizer_online.R, R/window_import_excel.R,
%   R/window_import_from_pkg.R, R/window_import_rdata.R,
%   R/window_import_rds.R, R/window_import_text_delim.R,
%   R/window_locale_set.R, R/window_model_select.R, R/window_rows_arrange.R,
%   R/window_rows_col_to_rownames.R, R/window_rows_filter.R,
%   R/window_rows_rm_duplicated.R, R/window_rows_rm_with_na.R,
%   R/window_rows_rowid_to_col.R, R/window_rows_rownames_to_col.R,
%   R/window_rows_slice.R, R/window_summary_count.R, R/window_summary_skim.R,
%   R/window_test_normality.R, R/window_variable_gather.R,
%   R/window_variable_mutate.R, R/window_variable_recode.R,
%   R/window_variable_rename.R, R/window_variable_select.R,
%   R/window_ws_obj_copy.R, R/window_ws_obj_delete.R, R/window_ws_obj_rename.R
\name{active_dataset}
\alias{active_dataset}
\alias{window_ci_mean__0}
\alias{window_ci_mean_diff}
\alias{window_ci_median}
\alias{window_ci_prop_multi}
\alias{window_ci_prop_binom}
\alias{window_ci_boot}
\alias{Menu-window-functions}
\alias{command_dataset_refresh}
\alias{command_dataset_view}
\alias{command_dataset_print}
\alias{command_dataset_print_as_df}
\alias{command_dataset_print_as_dt}
\alias{command_dataset_print_as_tibble}
\alias{command_dataset_class}
\alias{command_dataset_as_df}
\alias{command_dataset_as_tibble}
\alias{command_dataset_as_dt}
\alias{window_export_to_rds}
\alias{window_export_to_rdata}
\alias{window_export_to_excel_old}
\alias{window_export_to_textfile}
\alias{dataset_import_excel}
\alias{command_model_std_lm_coeffs}
\alias{command_model_get_class}
\alias{command_model_print}
\alias{command_model_summary}
\alias{command_model_glance}
\alias{command_model_tidy}
\alias{command_model_augment}
\alias{open_new_plots_window}
\alias{window_show_colors}
\alias{command_rownames}
\alias{command_rows_has_rownames}
\alias{command_rows_rm_empty_rows}
\alias{command_getwd}
\alias{command_setwd}
\alias{command_openwd}
\alias{command_get_file_info}
\alias{command_list_objects}
\alias{command_session_info_devtools}
\alias{command_session_info_utils}
\alias{command_dataset_dim}
\alias{command_glimpse}
\alias{summary_head_tail}
\alias{command_colnames}
\alias{command_clean_names}
\alias{command_all_chr_to_fctr}
\alias{copy_active_ds_to_clipboard}
\alias{open_online_tool}
\alias{window_online_geogebra_probability}
\alias{window_online_mvn}
\alias{window_online_gpower}
\alias{window_online_easyROC}
\alias{window_online_tdROC}
\alias{window_ci_mean}
\alias{window_dataset_bind_cols}
\alias{window_dataset_bind_rows}
\alias{window_dataset_class}
\alias{window_dataset_join}
\alias{window_dataset_print_as_md}
\alias{window_dataset_print_as_kable}
\alias{window_dataset_select}
\alias{window_export_to_excel}
\alias{window_factor_lvls_drop}
\alias{window_fct_lvls_reorder_manual}
\alias{window_image_digitizer_online}
\alias{window_import_excel}
\alias{command_list_datasets_in_pkgs}
\alias{window_import_from_pkg}
\alias{window_import_rdata}
\alias{window_import_rds}
\alias{window_import_text_delim}
\alias{window_import_text_delim0}
\alias{command_get_locale}
\alias{window_locale_set}
\alias{window_model_select}
\alias{window_rows_arrange}
\alias{window_rows_arrange_tmp}
\alias{window_rows_col_to_rownames0}
\alias{window_rows_col_to_rownames}
\alias{window_rows_filter0}
\alias{window_rows_filter}
\alias{window_rows_rm_duplicated}
\alias{window_rows_rm_with_na}
\alias{window_rows_rowid_to_col}
\alias{window_rows_rownames_to_col}
\alias{window_rows_slice}
\alias{window_summary_count}
\alias{window_summary_skim}
\alias{window_test_normality}
\alias{window_variable_gather}
\alias{window_variable_mutate0}
\alias{window_variable_mutate}
\alias{window_variable_recode0}
\alias{window_variable_rename}
\alias{window_variable_select0}
\alias{window_variable_select}
\alias{window_data_obj_copy}
\alias{window_data_obj_delete}
\alias{window_data_obj_rename}
\title{RcmdrPlugin.biostat functions for menus and windows.}
\usage{
active_dataset(dsname, flushModel = TRUE, flushDialogMemory = TRUE)

window_ci_mean__0()

window_ci_mean_diff()

window_ci_median()

window_ci_prop_multi()

window_ci_prop_binom()

window_ci_boot()

command_dataset_refresh()

command_dataset_view()

command_dataset_print()

command_dataset_print_as_df()

command_dataset_print_as_dt()

command_dataset_print_as_tibble()

command_dataset_class()

command_dataset_as_df()

command_dataset_as_tibble()

command_dataset_as_dt()

window_export_to_rds()

window_export_to_rdata()

window_export_to_excel_old()

window_export_to_textfile()

dataset_import_excel()

command_model_std_lm_coeffs()

command_model_get_class()

command_model_print()

command_model_summary()

command_model_glance()

command_model_tidy()

command_model_augment()

open_new_plots_window()

window_show_colors()

command_rownames()

command_rows_has_rownames()

command_rows_rm_empty_rows()

command_getwd()

command_setwd()

command_openwd()

command_get_file_info()

command_list_objects()

command_session_info_devtools()

command_session_info_utils()

command_dataset_dim()

command_glimpse()

summary_head_tail()

command_colnames()

command_clean_names()

command_all_chr_to_fctr()

copy_active_ds_to_clipboard()

open_online_tool(url = NULL, copy_to_clipboard = FALSE)

window_online_geogebra_probability()

window_online_mvn()

window_online_gpower()

window_online_easyROC()

window_online_tdROC()

window_online_tdROC()

window_ci_mean()

window_dataset_bind_cols()

window_dataset_bind_rows()

window_dataset_class()

window_dataset_join()

window_dataset_print_as_md()

window_dataset_print_as_kable()

window_dataset_select()

window_export_to_excel()

window_factor_lvls_drop()

window_fct_lvls_reorder_manual()

window_image_digitizer_online(variables)

window_import_excel()

command_list_datasets_in_pkgs()

window_import_from_pkg()

window_import_rdata()

window_import_rds()

window_import_text_delim()

window_import_text_delim0(init_location = "local")

command_get_locale()

window_locale_set()

window_model_select()

window_rows_arrange()

window_rows_arrange_tmp()

window_rows_col_to_rownames0(variables)

window_rows_col_to_rownames(new_dsname = NULL, init_conditions = NULL,
  incorrect_cond_msg = NULL)

window_rows_filter0(variables)

window_rows_filter(new_dsname = NULL, init_conditions = NULL,
  incorrect_cond_msg = NULL)

window_rows_rm_duplicated()

window_rows_rm_with_na()

window_rows_rowid_to_col()

window_rows_rownames_to_col()

window_rows_slice()

window_summary_count()

window_summary_skim()

window_test_normality()

window_variable_gather()

window_variable_mutate0(variables)

window_variable_mutate(var_name = NULL, init_express = NULL,
  incorrect_expr_msg = NULL)

window_variable_recode0()

window_variable_rename()

window_variable_select0(variables)

window_variable_select(new_dsname = NULL, incorrect_cond_msg = NULL)

window_data_obj_copy()

window_data_obj_delete()

window_data_obj_rename()
}
\description{
Functions that open Rcmdr menus and windows windows.
}
\keyword{internal}
