% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-menu.R, R/models-menu.R,
%   R/options-menu.R, R/plot-menu.R, R/summary-menu.R, R/variables-menu.R,
%   R/window_qqplot_biostat.R, R/window_test_normality.R
\name{Menu-winow-functions}
\alias{Menu-winow-functions}
\alias{command_dataset_refresh}
\alias{command_dataset_class}
\alias{command_dataset_as_df}
\alias{window_import_rds}
\alias{dataset_import_excel}
\alias{window_export_as_textfile}
\alias{window_export_as_rds}
\alias{command_std_lm_coeffs}
\alias{command_get_model_class}
\alias{command_load_packages}
\alias{command_getwd}
\alias{command_get_locale}
\alias{command_set_locale_lt}
\alias{command_set_locale_en}
\alias{open_new_plots_window}
\alias{summary_glimpse}
\alias{summary_Hmisc_describe}
\alias{summary_psych_describe}
\alias{summary_psych_describeBy}
\alias{command_all_chr_to_fctr}
\alias{window_qqplot_biostat}
\alias{window_normality_test}
\title{RcmdrPlugin.BioStat functions for menus and windows}
\usage{
command_dataset_refresh()

command_dataset_class()

command_dataset_as_df()

window_import_rds()

dataset_import_excel()

window_export_as_textfile()

window_export_as_rds()

command_std_lm_coeffs()

command_get_model_class()

command_load_packages()

command_getwd()

command_get_locale()

command_set_locale_lt()

command_set_locale_en()

open_new_plots_window()

summary_glimpse()

summary_Hmisc_describe()

summary_psych_describe()

summary_psych_describeBy()

command_all_chr_to_fctr()

window_qqplot_biostat()

window_normality_test()
}
\description{
Functions that open Rcmdr menus and windows windows
}
\keyword{internal}
