% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-functions.R, R/hepler-functions.R,
%   R/plot-functions.R, R/summary-functions.R, R/variable-functions.R,
%   R/window_qqplot_biostat.R, R/window_test_normality.R
\name{Menu-winow-functions}
\alias{Menu-winow-functions}
\alias{dataset_refresh}
\alias{dataset_class}
\alias{dataset_as_df}
\alias{dataset_import_rds}
\alias{window_export_as_textfile}
\alias{window_export_as_rds}
\alias{load_packages_command}
\alias{new_plots_window}
\alias{summary_glimpse}
\alias{summary_Hmisc_describe}
\alias{summary_psych_describe}
\alias{command_all_chr_to_fctr}
\alias{window_qqplot_biostat}
\alias{window_normality_test}
\title{RcmdrPlugin.BioStat functions for menus and windows}
\usage{
dataset_refresh()

dataset_class()

dataset_as_df()

dataset_import_rds()

window_export_as_textfile()

window_export_as_rds()

load_packages_command()

new_plots_window()

summary_glimpse()

summary_Hmisc_describe()

summary_psych_describe()

command_all_chr_to_fctr()

window_qqplot_biostat()

window_normality_test()
}
\description{
Functions that open Rcmdr menus and windows windows
}
\keyword{internal}
