% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R, R/menu-conf_interval.R,
%   R/menu-data.R, R/menu-export.R, R/menu-import.R, R/menu-models.R,
%   R/menu-plots.R, R/menu-settings.R, R/menu-summary.R, R/menu-variables.R,
%   R/window_anova_kw_mood_tests.R, R/window_do_summary.R,
%   R/window_qqplot_biostat.R, R/window_test_normality.R, R/window_tmp.R
\name{function_not_implemented}
\alias{function_not_implemented}
\alias{window_ci_mean}
\alias{window_ci_prop}
\alias{window_ci_boot}
\alias{Menu-winow-functions}
\alias{command_dataset_refresh}
\alias{command_dataset_class}
\alias{command_dataset_as_df}
\alias{window_export_as_textfile}
\alias{window_export_as_rds}
\alias{window_import_rds}
\alias{dataset_import_excel}
\alias{command_std_lm_coeffs}
\alias{command_get_model_class}
\alias{open_new_plots_window}
\alias{window_histogram}
\alias{window_density_plot}
\alias{window_plot_dot_err}
\alias{window_plot_bar_err}
\alias{command_load_packages}
\alias{command_getwd}
\alias{command_get_locale}
\alias{command_set_locale_lt}
\alias{command_set_locale_en}
\alias{summary_glimpse}
\alias{summary_head_tail}
\alias{summary_Hmisc_describe}
\alias{summary_psych_describe}
\alias{summary_psych_describeBy}
\alias{command_all_chr_to_fctr}
\alias{window_anova_kw_mood_tests}
\alias{window_do_summary}
\alias{window_qqplot_biostat}
\alias{window_normality_test}
\alias{window_tmp}
\title{RcmdrPlugin.BioStat functions for menus and windows}
\usage{
function_not_implemented(x = NULL)

window_ci_mean()

window_ci_prop()

window_ci_boot()

command_dataset_refresh()

command_dataset_class()

command_dataset_as_df()

window_export_as_textfile()

window_export_as_rds()

window_import_rds()

dataset_import_excel()

command_std_lm_coeffs()

command_get_model_class()

open_new_plots_window()

window_histogram()

window_density_plot()

window_plot_dot_err()

window_plot_bar_err()

command_load_packages()

command_getwd()

command_get_locale()

command_set_locale_lt()

command_set_locale_en()

summary_glimpse()

summary_head_tail()

summary_Hmisc_describe()

summary_psych_describe()

summary_psych_describeBy()

command_all_chr_to_fctr()

window_anova_kw_mood_tests()

window_do_summary()

window_qqplot_biostat()

window_normality_test()

window_tmp()
}
\description{
Functions that open Rcmdr menus and windows windows
}
\keyword{internal}
