% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hepler-functions.R
\name{modelClassP}
\alias{modelClassP}
\title{Chech the class of the active model in Rcmdr}
\usage{
modelClassP(class_)
}
\arguments{
\item{class_}{(string) a character vector of length 1. The name of class.}
}
\description{
Chech the class of the active model in Rcmdr
}
\keyword{internal}
