% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gpie.R
\name{windowPie}
\alias{windowPie}
\title{Wrapper Function of Pie Charts Subclass}
\usage{
windowPie()
}
\description{
\code{windowPie} function is a wrapper function of \code{gpie} class for the R-commander menu bar.
}
\keyword{hplot}
