% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package.R
\docType{package}
\name{RcmdrPlugin.BioStat}
\alias{RcmdrPlugin.BioStat}
\alias{RcmdrPlugin.BioStat-package}
\title{An R Commander Plugin for Basic (Bio)Statistical Routines}
\description{
\code{RcmdrPlugin.BioStat} is an \code{Rcmdr} plugin for functions in package
\code{BioStat} and other common (bio)statistical routines.
}
