% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package.R
\docType{package}
\name{RcmdrPlugin.biostat}
\alias{RcmdrPlugin.biostat}
\alias{RcmdrPlugin.biostat-package}
\title{An R Commander Plug-in for Basic Data Analysis Tasks}
\description{
\code{RcmdrPlugin.biostat} is an \code{Rcmdr} plugin for functions in package
\code{biostat} and other common (bio)statistical routines.
}
