% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package.R
\docType{package}
\name{RcmdrPlugin.biostat}
\alias{RcmdrPlugin.biostat}
\title{R Commander Plug-in for Basic Data Management and Analysis Tasks}
\description{
\pkg{RcmdrPlugin.biostat} is an
\href{\pkg{Rcmdr}}{https://CRAN.R-project.org/package=Rcmdr}
plug-in for the most common data wrangling, visualisation and analysis tasks
using "tidyverse" family functions as well as functions from other packages.
}
