% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package.R
\docType{package}
\name{RcmdrPlugin.BioStat}
\alias{RcmdrPlugin.BioStat}
\alias{RcmdrPlugin.BioStat-package}
\title{A Set of RStudio Addins}
\description{
\code{RcmdrPlugin.BioStat} is an \code{Rcmdr} plugin for package
\code{BioStat}.
}
