% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_elements.R
\name{TclTk-helper-functions}
\alias{TclTk-helper-functions}
\alias{bs_label_b}
\alias{bs_label}
\alias{labeled_frame}
\alias{tk_title}
\alias{bs_title}
\alias{tcl_get_parent}
\alias{tcl_get_children}
\alias{tcl_get_siblings}
\title{Helper functions for Tcl/Tk.}
\usage{
bs_label_b(..., fg = Rcmdr::getRcmdr("title.color"))

bs_label(..., fg = NULL)

labeled_frame(parent, label = NULL, ...)

tk_title(parent = top, text = "xxx_title", pady = c(5, 9),
  font = tkfont.create(weight = "bold", size = 9),
  fg = Rcmdr::getRcmdr("title.color"), ...)

bs_title(parent = top, text, pady = c(5, 9),
  font = tkfont.create(weight = "bold", size = 9), ...)

tcl_get_parent(obj)

tcl_get_children(obj)

tcl_get_siblings(obj)
}
\description{
Helper functions for Tcl/Tk.
}
\keyword{internal}
