% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_elements.R
\name{TclTk-labels}
\alias{TclTk-labels}
\alias{bs_label}
\alias{bs_label_title}
\alias{bs_label_b}
\alias{tk_label_blue}
\title{Tcl/Tk labels}
\usage{
bs_label(..., fg = NULL)

bs_label_title(parent, text, ..., fg = Rcmdr::getRcmdr("title.color"))

bs_label_b(...)

tk_label_blue(...)
}
\arguments{
\item{...}{Other arguments to pass to \code{tcltk2::tk2label()}.}

\item{fg}{(character) Foreground color.}

\item{parent}{(\code{"tkwin"} object) Parant Tcl/Tk frame.}

\item{text}{(character) Label text.}
}
\description{
Functions, tat create Tcl/Tk labels.
}
\seealso{
\code{tcltk2::tk2label()}, \code{Rcmdr::labelRcmdr()}, \code{tcltk::ttklabel()}.
}
