% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk_commands.R
\name{tk_activate}
\alias{tk_activate}
\alias{tk_activate.default}
\alias{tk_normalize.default}
\alias{tk_read_only.default}
\alias{tk_disable.default}
\title{Activate/Disable TK Objects}
\usage{
\method{tk_activate}{default}(obj, ...)

\method{tk_normalize}{default}(obj, ...)

\method{tk_read_only}{default}(obj, ...)

\method{tk_disable}{default}(obj, ...)
}
\arguments{
\item{obj}{TK object.}

\item{...}{other options to be passed to \code{tkconfigure}.}
}
\description{
Modify state of tk objects.
}
\examples{
# tk_activate(obj)
# tk_disable(obj)

}
\keyword{internal}
