% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcktk_commands.R
\name{tk_activate}
\alias{tk_activate}
\alias{tk_activate}
\alias{tk_disable}
\alias{tk_activate}
\alias{tk_normalize}
\title{Activate/Disable TK Objects}
\usage{
tk_activate(tk_obj, ...)

tk_disable(tk_obj, ...)

tk_normalize(tk_obj, ...)
}
\arguments{
\item{tk_obj}{TK object.}

\item{...}{other options to be passed to \code{tkconfigure}.}
}
\description{
Modify state of tk objects.
}
\examples{
# tk_activate(tk_obj)
# tk_disable(tk_obj)

}
\keyword{internal}
