% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{variables_fct}
\alias{variables_fct}
\title{True factor variables names in active dataset}
\usage{
variables_fct()
}
\description{
True factor variables names in active dataset
}
\keyword{internal}
