% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_log_transform.R
\name{window_log_transform}
\alias{window_log_transform}
\title{window_log_transform}
\usage{
window_log_transform()
}
\description{
window_log_transform
}
\seealso{
Other transformations: \code{\link{window_z_transform_2}},
  \code{\link{window_z_transform}}
}
