% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_variable_convert_type.R
\name{window_variable_convert_type}
\alias{window_variable_convert_type}
\title{Rcmdr windows for variable class conversion}
\usage{
window_variable_convert_type()
}
\description{
Rcmdr windows for variable class conversion
}
\concept{conversion}
\keyword{internal}
