% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_z_transform.R
\name{window_z_transform}
\alias{window_z_transform}
\title{Rcmdr window for z transformation}
\usage{
window_z_transform()
}
\description{
Rcmdr window for z transformation
}
\seealso{
Other transformations: \code{\link{window_log_transform}},
  \code{\link{window_z_transform_2}}
}
\keyword{internal}
